% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{apply_portfolio_multiplier}
\alias{apply_portfolio_multiplier}
\title{Apply Portfolio Multiplier To Positions}
\usage{
apply_portfolio_multiplier(
  t,
  prices,
  modified_target_pos_ccy,
  portfolio_multiplier_value,
  t_last_position_entry,
  latest_trade_direction,
  position_tables,
  config
)
}
\arguments{
\item{portfolio_multiplier_value}{Portfolio multiplier value.}

\item{position_tables}{Position tables}

\item{...}{}
}
\value{
List of updated position tables
}
\description{
Multiply position sizes by portfolio multiplier and update position tables.
}
