% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{apply_rule}
\alias{apply_rule}
\title{Apply Opening Or Closing Rule To Data}
\usage{
apply_rule(algo, signal_table, position_table, t)
}
\arguments{
\item{t}{}
}
\description{
Apply a single rule to a single instrument.

The \code{algo} list contains a rule function and data for a single instrument.
}
