% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{backfill_columns}
\alias{backfill_columns}
\title{Backfill specific columns in data frame}
\usage{
backfill_columns(t, table, column_names, value = NA)
}
\arguments{
\item{t}{Time index of the row succeeding the last row to be backfilled}

\item{table}{A data frame}

\item{column_names}{A character vector of column names}

\item{value}{Value to fill with. Default is NA.}
}
\value{
A list of data frame columns
}
\description{
Backfill specific columns in data frame
}
