% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark}
\usage{
benchmark(n = 1, expr, envir = parent.frame())
}
\arguments{
\item{n}{Number of test runs per function.}

\item{expr}{Function call as character string. E.g. \code{"my_function(df)"}.}

\item{envir}{Environment in wich to evaluate the function. Typically
\code{benchmark()} is called from \code{run_benchmark()} so the tested function will
be evaluated in the environment of \code{run_benchmark()}.}
}
\description{
Run benchmark test on a single function.

Execute \code{expr} \code{n} times.

Explicit garbage collection, then execute \code{expr} \code{n} times w/o explicit gc,
finally return timings.

A progress bar shows progress for each individual run, when running time is
more than 2 secs and expected remaining time is more than 2 secs.
}
