% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{buffer_position}
\alias{buffer_position}
\title{Buffer Position Modifier}
\usage{
buffer_position(t, position_size_ccy, rel_buffer_size, position_table)
}
\arguments{
\item{t}{Time index.}

\item{position_size_ccy}{Position size in account currency.}

\item{rel_buffer_size}{Relative buffer size.}

\item{position_table}{Position table.}

\item{t_last_position_entry}{Time index of last position entry.}
}
\description{
Buffer Position Modifier
}
\details{
Keep last position size if
\deqn{\dfrac{|\text{position}_{t} - \text{position}_{t - 1}|}{\text{position}_{t - 1}} > \text{Relative buffer threshold}}
}
