% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{calculate_equal_inst_weights}
\alias{calculate_equal_inst_weights}
\title{Calculate Equal Instrument Weights From Parsed Algos}
\usage{
calculate_equal_inst_weights(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\value{
Vector of weights.
}
\description{
Calculates a vector of weights, one weight for each instrument. The sum of
all weights is 1.
}
