% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{calculate_equal_signal_weight}
\alias{calculate_equal_signal_weight}
\title{Calculate Equal Signal Weight}
\usage{
calculate_equal_signal_weight(algos)
}
\arguments{
\item{algos}{}
}
\description{
Calculate Equal Signal Weight
}
