% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{calculate_equal_signal_weights}
\alias{calculate_equal_signal_weights}
\title{Calculate Equal Signal Weights For All Algos}
\usage{
calculate_equal_signal_weights(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos}
}
\value{
Vector of weights.
}
\description{
First calculates the equal weights for each instrument. Then outputs a vector
of weights, one weight for each parsed algo. The sum of all weights for
each subsystem is 1.
}
