% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{calculate_portfolio_multiplier}
\alias{calculate_portfolio_multiplier}
\title{Calculate Portfolio Multiplier}
\usage{
calculate_portfolio_multiplier(portfolio_multiplier, mul_variable_param_vals)
}
\arguments{
\item{portfolio_multiplier}{List of portfolio multiplier specifications.}

\item{mul_variable_param_vals}{Variable parameter values for portfolio
multiplier.}
}
\value{
List of portfolio multiplier outputs. First element is the multiplier
value.
}
\description{
Calculate Portfolio Multiplier
}
