% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{calculate_signal_weights}
\alias{calculate_signal_weights}
\title{Calculate Signal Weights For All Algos}
\usage{
calculate_signal_weights(algos, method = "equal")
}
\arguments{
\item{algos}{List of all exoanded algos.}

\item{method}{Method for calculating weights.}
}
\value{
Numeric vector of weights
}
\description{
Calculate signal weights for all algos.

Within each subsystem weights must sum to one. Note, the sum of signal
weights for all algos need not sum to one.
}
