% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{calculate_stop_loss_signal}
\alias{calculate_stop_loss_signal}
\title{Stop loss signal}
\usage{
calculate_stop_loss_signal(
  prices,
  t,
  t_trade_open,
  stop_loss_gap,
  direction = 0,
  rnd = FALSE
)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{t}{Time index.}

\item{t_trade_open}{Time index of the time when trade is opened.}

\item{stop_loss_gap}{Stop loss gap.}

\item{direction}{Is current trade long or short? 1 for long, -1 for short.}

\item{rnd}{If TRUE, add small random amount to stop loss level. Negative if
short.}
}
\value{
-1, 0 or 1.
}
\description{
Calculate stop loss signal.
}
