% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{calculate_subsystem_position}
\alias{calculate_subsystem_position}
\title{Calculate Subsystem Position}
\usage{
calculate_subsystem_position(combined_signal, required_leverage_factor)
}
\arguments{
\item{combined_signal}{Typically normalized and clamped}

\item{required_leverage_factor}{Required leverage factor}
}
\value{
Number of contracts
}
\description{
Calculate Subsystem Position
}
