% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cap_signal}
\alias{cap_signal}
\title{Cap}
\usage{
cap_signal(signal, min_signal, max_signal)
}
\arguments{
\item{signal_max}{}
}
\description{
Cap a signal between a min and a max value.

If input value is NA, output will be NA too.
}
