% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp_matrix}
\alias{clamp_matrix}
\title{Clamp Matrix Elements At Upper And Lower Limit}
\usage{
clamp_matrix(input_matrix, min_signal = -Inf, max_signal = Inf)
}
\arguments{
\item{input_matrix}{Input matric}

\item{min_signal}{Minimum signal value}

\item{max_signal}{Maximum signal value}
}
\description{
For each element in the matrix limit any signal value below \code{min_signal} to
\code{min_signal} and any value above \code{max_signal} to \code{max_signal}.
}
