% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp_matrix_lower}
\alias{clamp_matrix_lower}
\title{Clamp Matrix Elements At Lower Limit}
\usage{
clamp_matrix_lower(input_matrix, min_signal)
}
\arguments{
\item{input_matrix}{Input matrix}

\item{min_signal}{Minimum signal value}
}
\description{
For each element in the matrix limit any signal value below \code{min_signal} to
\code{min_signal}.
}
