% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp_matrix_upper}
\alias{clamp_matrix_upper}
\title{Clamp Matrix Elements At Upper Limit}
\usage{
clamp_matrix_upper(input_matrix, max_signal)
}
\arguments{
\item{input_matrix}{Input matrix}

\item{max_signal}{Maximum signal}
}
\description{
For each element in the matrix limit any signal value above \code{max_signal} to
\code{max_signal}.
}
