% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp_signal}
\alias{clamp_signal}
\title{Clamp Signal At Upper And Lower Limits}
\usage{
clamp_signal(signal, min_signal = -Inf, max_signal = Inf)
}
\arguments{
\item{signal}{Single number or numeric vector}

\item{min_signal}{Minimum signal value}

\item{max_signal}{Maximum signal value}
}
\description{
Clamp a signal between a min and a max value.

If input value is NA, output will be NA too.
}
