% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp_signal_lower}
\alias{clamp_signal_lower}
\title{Clamp Single Signal Value At Lower Limit}
\usage{
clamp_signal_lower(signal, min_signal = -Inf)
}
\arguments{
\item{signal}{Single signal value}

\item{min_signal}{Minimum signal value}
}
\value{
Number
}
\description{
Limit any signal value below \code{min_signal} to \code{min_signal}.
}
