% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp_signal_upper}
\alias{clamp_signal_upper}
\title{Clamp Signal At Upper Limit}
\usage{
clamp_signal_upper(signal, max_signal = Inf)
}
\arguments{
\item{signal}{Signal}

\item{max_signal}{Maximum signal value}
}
\description{
Limit any signal value above \code{max_signal} to \code{max_signal}.
}
