% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{combine_portfolio_multipliers}
\alias{combine_portfolio_multipliers}
\title{Combine Portfolio Multipliers}
\usage{
combine_portfolio_multipliers(
  portfolio_multipliers,
  system_vars,
  combi_method = "min"
)
}
\arguments{
\item{portfolio_multipliers}{List of multipliers.}

\item{combi_method}{Combination method.}

\item{mul_variable_param_vals}{Multiplier variable parameter values.}
}
\value{
List of combined portfolio multiplier output. First element is the
combined multiplier values, as specified by the \code{combi_method}. Remaining
elements are additional output from multiplier functions.
}
\description{
Calculates the combined output from multiple portfolio multiplier functions.
Multiplier values are combined according to the \code{combi_method}.
}
\details{
Each portfolio multiplier in the portfolio multipliers list contains a
portfolio multiplier specification list:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{portfolio_multipliers = list(
  list(
    modifier_name = <multiplier-name-1>,
    modifier_function = <multiplier-function-1>,
    <fixed-params-1_1>,
    <fixed-params-1_2>
  ),
  list(
    modifier_name = <multiplier-name-2>,
    modifier_function = <multiplier-function-2>,
    <fixed-params-2_1>,
    <fixed-params-2_2>
  )
)
}\if{html}{\out{</div>}}

Portfolio multiplier functions are combined according to the method specified
in \code{my_system$config$portfolio_multiplier$combi_method}. The default is to
use the minimum value. Valid combination methods are \code{min}, \code{max}, \code{add}
and \code{multiply}.

Each input portfolio multiplier specification list specifies a multiplier
name, a multiplier function and any number of fixed parameters. The parser
automatically determines which parameters of each multiplier function are
fixed parameters provided by the user, and which are variable parameters
provided by the system.

First element in each input \code{portfolio_multipliers[[i]]} list is the
multiplier name, second element is the multiplier function definition
itself. Remaining elements are the fixed params.

Note: You may provide any name as multiplier name. If the multiplier
function is assigned to a variable, you may put this variable as the
function definition. So the function name doesn't have to match the
multiplier name. The multiplier name is just a label for your convenience.

Note: Variable params are not provided by user, i.e. do not appear
in multiplier list.
}
