% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_binary_class_label}
\alias{convert_binary_class_label}
\title{Convert Binary Class Label}
\usage{
convert_binary_class_label(input, from_format = "binary", to_format = "sign")
}
\arguments{
\item{input}{0 or 1 for "binary" input. -1 or 1 for "sign" input.}

\item{from_format}{Format to convert from.}

\item{to_format}{Format to convert to.}
}
\value{
Number
}
\description{
Converts binary label from "binary" format to "sign" format.
}
\details{
\itemize{
\item "binary" format: \eqn{\{0, 1\}}
\item "sign" format: \eqn{\{-1, 1\}}
\item "auto format": Automatically converts from "binary" to "sign" or vice
versa. This is a faster than the other options, but doesn't check
validity of the input. WARNING: Nonsensical input will produce nonsensical
output. E.g. input 2 will produce output -3. "to_format" is ignored, if
"from_format" is "auto".
}

If output format is same as input format, it will not fail, just return the
input unchanged.
}
\examples{
convert_binary_class_label(0) # -1
convert_binary_class_label(1) # 1
convert_binary_class_label(-1, "sign", "binary") # 0
convert_binary_class_label(0, "sign", "binary") # 0 not valid input
}
