% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_function_from_string}
\alias{eval_function_from_string}
\title{Evaluate Function From String}
\usage{
eval_function_from_string(function_name)
}
\arguments{
\item{function_name}{Function name as a character string.}
}
\value{
Function
}
\description{
Evaluate a function given a function name as a character string.
}
