% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{eval_inst}
\alias{eval_inst}
\title{Evaluate Instrument Data Set}
\usage{
eval_inst(instrument_name)
}
\arguments{
\item{rule_name}{Instrument data set name as a character string.}
}
\value{
Data frame
}
\description{
Evaluate an instrument data set given a name as a character string.
}
