% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{eval_rule}
\alias{eval_rule}
\title{Evaluate Rule Function}
\usage{
eval_rule(rule_name)
}
\arguments{
\item{rule_name}{Rule function name as a character string.}
}
\value{
Function
}
\description{
Evaluate a rule function given a function name as a character string.
}
