% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{ew_cor_matrix}
\alias{ew_cor_matrix}
\title{Exponential Weighted Correlation Matrix}
\usage{
ew_cor_matrix(data, lambda = NA, lookback = NA)
}
\arguments{
\item{data}{Dataframe.}

\item{lookback}{Integer. Lookback window length.}
}
\value{
Matrix
}
\description{
Exponentially Weighted Moving-Average estimate of the correlation matrix.
See Tsay: Analysis Of Financial Time Series (3rd Ed., 10.1, p. 507)
}
