% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{ew_cov_matrix}
\alias{ew_cov_matrix}
\title{Exponential Wrighted Covariance Matrix}
\usage{
ew_cov_matrix(data, lookback = NA)
}
\arguments{
\item{data}{Dataframe.}

\item{lookback}{Integer. Lookback window length.}
}
\value{
Matrix
}
\description{
Exponentially Weighted Moving-Average estimate of the covariance matrix.
See Tsay: Analysis Of Financial Time Series (3rd Ed., 10.1, p. 507)
}
