% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exists_and_not_null}
\alias{exists_and_not_null}
\title{Check That Object Exists And Is Not Null}
\usage{
exists_and_not_null(x, warn_or_stop = "warn")
}
\arguments{
\item{x}{Object name as character string.}

\item{warn_or_stop}{Option to throw warning (\code{"warn"}) or stop (\code{"stop"}).}
}
\value{
Boolean
}
\description{
Validate that object exists and is not null or empty. Returns \code{FALSE} if \code{x}
does not exist, is \code{NULL} or is empty. Otherwise returns \code{TRUE}.

Main use case is to check that a vector, list or data frame is not empty.
While an empty vector will also be \code{NULL}, an empty list is not \code{NULL}.
So this validator function checks both, given that the input object exists.
}
\details{
Validation should return \code{FALSE} for data frames with all empty columns as
well. The validation function is using the length of the object to
determine if it is empty. Eventhough the length of a data frame is the
number of columns (not rows!), empty columns do not count in the number of
columns of a data frame.

Note, in most situations this validation should be redundant, as R already
will let us know if an object is missing or \code{NULL}. E.g. checking that an
input object exists at the top of a function definition should not be
necessary, because R won't execute a function if an input object doesn't
exist (it could still be \code{NULL}, though). But we might like to validate an
object before we pass it to a function. The validation function allows us
to check before R starts complaining on it's own.

Also note that the function throwing any errors will be
\code{exists_and_not_null()}, which may not be helpful.
}
