% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_account_pandl}
\alias{f_account_pandl}
\title{Calculate Account Profit/Loss}
\usage{
f_account_pandl(position_tables, t)
}
\arguments{
\item{position_tables}{List of all position tables.}

\item{t}{Time index ("now")}
}
\value{
List of profits/losses
}
\description{
Calculate account profit/loss.

Calculate account profit and loss at time \eqn{t}.
}
