% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_account_value}
\alias{f_account_value}
\title{Calculate Account Value}
\usage{
f_account_value(position_tables, cash, t)
}
\arguments{
\item{position_tables}{List of all position tables.}

\item{cash}{Amount in cash account.}

\item{t}{Time index ("now").}
}
\value{
Single number in units of price
}
\description{
Current value of all open positions plus cash.
}
