% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_average}
\alias{f_average}
\title{Calculate Average}
\usage{
f_average(x, method = "simple", ...)
}
\arguments{
\item{x}{Vector.}

\item{method}{Calculation method.
\itemize{
\item \code{"simple"} Simple mean.
\itemize{
\item @seealso \code{\link[=mean]{mean()}}
}
\item \code{"ewa"} Exponentially Weighted Average.
\itemize{
\item Additional parameters: \code{lambda}(number), \code{lookback} (integer).
\item @seealso \code{\link[=f_ewa]{f_ewa()}}
}
}}

\item{...}{}
}
\value{
Vector of moving standard deviations.
}
\description{
Calculate average using a specified method.
}
\examples{
f_average(1:10, method = "ewa", lambda = 0.8, lookback = 4L)
}
