% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_capital}
\alias{f_capital}
\title{Calculate Trading Capital}
\usage{
f_capital(system_account_table, account_pandl, t)
}
\arguments{
\item{account_pandl}{Total account profit/loss at time \eqn{t}.}

\item{t}{}

\item{account_table}{Account table.}
}
\value{
Single amount in units of price
}
\description{
The trading capital is the initial trading capital plus the sum of all
profits/losses. Profits/losses are mark to market, not just realized.
}
