% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-alpha.R
\name{f_close_trade_stop_loss}
\alias{f_close_trade_stop_loss}
\title{Close Position Based On Stop Loss}
\usage{
f_close_trade_stop_loss()
}
\description{
Close position if stop loss level was breached yesterday.
Close position even if price has recovered.
}
