% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_cor_mat}
\alias{f_cor_mat}
\title{Correlation Matrix}
\usage{
f_cor_mat(data, method = "Pearson", ...)
}
\arguments{
\item{data}{Data frame.}

\item{method}{Method of correlation calculation as character string.
\itemize{
\item \code{"Pearson"}
\item \code{"ewa"} Exponentially Weighted Correlation. @seealso \code{\link[=f_ewa_cor_mat]{f_ewa_cor_mat()}}
\itemize{
\item Additional parameters: \code{lambda} (number), \code{lookback} (integer)
}
}}

\item{...}{}
}
\value{
Correlation matrix
}
\description{
Calculate correlation of columns in a data frame.
}
\details{
For \code{method = "ewa"}: If \code{L == NA}, \code{L} will be set to the number of rows of
the data. If \code{lambda == NA}, \code{lambda} will be set to
\code{1 - (2 / (1 + L))}.
}
