% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_ewa}
\alias{f_ewa}
\title{Calculate Exponentially Weighted Average}
\usage{
f_ewa(x, lambda = NA, lookback = NA)
}
\arguments{
\item{x}{Vector. Top to bottom: Oldest to newest.}

\item{lambda}{Smoothing parameter. If no lambda is provided, set lambda to
\eqn{1 - (2 / (1 + L))}.}

\item{lookback}{Lookback window length as positive integer. This is the tail
of the vector. If no \code{lookback} is provided, the entire \eqn{x} vector will
be used.}
}
\value{
Single exponentially weighted average value
}
\description{
Calculate EWA of vector \eqn{x}.
}
\details{
\deqn{E[X_t | X_{t-1}] = \frac{1}{\sum_{i=0}^{t-2}\lambda^i}\sum_{j=1}^{t-1}\lambda^{j-1} X_{t-j}}

Note: According to the formula, if the length of the lookback window is \eqn{L},
the range of the lookback window is \eqn{[t-L, t-1]}.
In practice we may want to calculate an average of prices in a window
from day \eqn{t - L + 1} to day \eqn{t} as soon as (the same day as) we have the
price for day \eqn{t}, rather than the next day as the formula would imply.
In this case the last  observation in the window would have the same time
index as the calculated average. \code{f_ewa()} does not involve a time index,
but these considerations come in to play when  applying \code{f_ewa()} with
\code{rolling_window()} or \code{rolling_window_step()}.
}
\examples{
f_ewa(1:10, lambda = 0.8, lookback = 3L)

}
\references{
Tsay: Analysis Of Financial Time Series (3rd Ed., 10.1, p. 507)
}
