% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_ewa_cor}
\alias{f_ewa_cor}
\title{Exponential Weighted Correlation}
\usage{
f_ewa_cor(x, y, lambda = NA, lookback = NA)
}
\arguments{
\item{x}{Vector.}

\item{y}{Vector.}

\item{lambda}{Smoothing parameter.}

\item{lookback}{Integer. Lookback window length. If no \code{lookback} is
provided, the entire \eqn{x} and \eqn{y} vectors will be used.}
}
\value{
Number
}
\description{
Exponential Weighted Correlation
}
\details{
\deqn{s^2(X) = \frac{1}{n - 1}\sum_i^n (X_i - \mu_X)^2}
\deqn{\text{COV}(X, Y) = \frac{1}{n - 1} \sum_i^n ((X_i - \mu_X)(Y_i - \mu_Y))}
\deqn{\text{COR}(X, Y) = \frac{1}{n - 1} \frac{\sum_i^n ((X_i - \mu_X)(Y_i - \mu_Y))}{s(X) s(Y)}}
}
\references{
Tsay: Analysis Of Financial Time Series (3rd Ed., 10.1, p. 507)
}
