% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_ewa_cor_mat}
\alias{f_ewa_cor_mat}
\title{Exponential Weighted Correlation Matrix}
\usage{
f_ewa_cor_mat(data, lambda = NA, lookback = NA)
}
\arguments{
\item{data}{Dataframe.}

\item{lambda}{Smoothing parameter. If \code{NA}}

\item{lookback}{Integer. Lookback window length.}
}
\value{
Matrix
}
\description{
Exponentially Weighted Moving-Average estimate of the correlation matrix.
}
\references{
Tsay: Analysis Of Financial Time Series (3rd Ed., 10.1, p. 507)
}
