% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_ewa_cov_mat}
\alias{f_ewa_cov_mat}
\title{Exponential Weighted Covariance Matrix}
\usage{
f_ewa_cov_mat(data, lambda = NA, lookback = NA)
}
\arguments{
\item{data}{Dataframe.}

\item{lambda}{Smoothing parameter.}

\item{lookback}{Integer. Lookback window length.}
}
\value{
Matrix
}
\description{
Exponentially Weighted Moving-Average estimate of the covariance matrix.
}
\references{
Tsay: Analysis Of Financial Time Series (3rd Ed., 10.1, p. 507)
}
\seealso{
\code{\link[=f_ewa]{f_ewa()}}
}
