% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_ewa_sd}
\alias{f_ewa_sd}
\title{Exponential Weighted Standard Deviation}
\usage{
f_ewa_sd(x, lambda = NA, lookback = NA)
}
\arguments{
\item{x}{Input vector.}

\item{lambda}{Smoothing parameter.}

\item{lookback}{Length of lookback window.}
}
\value{
Single value
}
\description{
Exponential Weighted Standard Deviation
}
\examples{
f_sd(1:10, lambda = 0.8, lookback = 3L)
}
\seealso{
\code{\link[=f_ewa]{f_ewa()}}
}
