% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_IDM}
\alias{f_IDM}
\title{Instrument Diversification Multiplier}
\usage{
f_IDM(correlations, weights)
}
\arguments{
\item{correlations}{Correlations matrix}

\item{weights}{Vector of weights}
}
\value{
Scalar
}
\description{
$$\text{IDM} = \frac{1}{\sqrt{W^{T}HW}}$$
Where $H$ is the correlation matrix for returns, and $W$ is the vector
of instrument weights summing to one.
}
