% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_indiv_normalization_factor}
\alias{f_indiv_normalization_factor}
\title{Calculate Normalization Factor From Signal Itself}
\usage{
f_indiv_normalization_factor(raw_signal_vector, target = 1)
}
\arguments{
\item{target}{The target expected value of the signal scaled by the
normalization factor.}

\item{raw_signal}{A vector of signal values on which the normalization is
based.}
}
\description{
Calculate normalization factor based on the mean absolute value of past
values of the signal itself.

See update_normalization_factors() for better options.

Each raw signal is scaled by a normalization factor. This normalization
factor is the \emph{required leverage target}.
f_indiv_normalization_factor() calculates the normalization factor based
on a vector of (previous) raw signal values.
The normalization factor is

\deqn{| \text{target} / E(\text{abs(signal)}) |}

As a default the expected absolute value of the scaled values will be 1.
}
