% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_inst_div_mul}
\alias{f_inst_div_mul}
\title{Instrument Diversification Multiplier (IDM)}
\usage{
f_inst_div_mul(
  inst_ret_cor_mat,
  instrument_weights,
  min_cor = 0,
  max_idm = 2.5
)
}
\arguments{
\item{inst_ret_cor_mat}{Subsystem returns correlations matrix.}

\item{instrument_weights}{Vector of weights.}

\item{min_cor}{Minimum value for each element in \code{inst_ret_cor_mat}
matrix.}

\item{max_idm}{Maximum output value of IDM.}
}
\value{
Scalar
}
\description{
\deqn{\text{IDM} = \frac{1}{\sqrt{W^{T}HW}}}
Where $H$ is the correlation matrix for subsystem returns in percentage
terms, and $W$ is the vector of instrument weights summing to one.
}
