% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_instr_risk}
\alias{f_instr_risk}
\title{Instrument risk, annualized}
\usage{
f_instr_risk(prices, t, window_length = NA)
}
\arguments{
\item{prices}{A time series of prices in currency. Newest first.
Top to bottom: Newer to older.}

\item{t}{Time index.}

\item{window_length}{Window length. Includes all if \code{window_length} is NA.}
}
\value{
Standard deviation
}
\description{
Standard deviations of returns.
}
\details{
The square root of the number of business days per year is hard coded:
sqrt(252) = 15.87451
252 is the number of business days in a year.
(We could also cound off to 16, and we would probably be fine.)
}
