% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-portfolio.R
\name{f_leverage_factor}
\alias{f_leverage_factor}
\title{Calculate Leverage Factor}
\usage{
f_leverage_factor(cash = NA, invested = NA, borrowed = NA)
}
\arguments{
\item{cash}{Amount of liquid available.}

\item{invested}{Amount to invest in an instrument.}

\item{borrowed}{Amount needed to borrow.}
}
\description{
Calculates leverage factor from any two provided valid parameters. At least
two valid parameters must be provided.
}
