% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-rules.R
\name{f_ma_vector}
\alias{f_ma_vector}
\title{Calculate Moving Average Vector}
\usage{
f_ma_vector(prices, n)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom: Newer to older.}

\item{n}{Moving Average lookback period.}
}
\value{
A vector of prices.
}
\description{
Calculates moving average for each row in a price series data frame.
N (length of price vector) smaller than or equal to n is accepted.
This will not be the desired moving average, but will also not fail.
}
