% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_min_capital}
\alias{f_min_capital}
\title{Minimum capital}
\usage{
f_min_capital(min_exposure, instr_risk, instrument_risk_target)
}
\arguments{
\item{instrument_risk_target}{}
}
\description{
Minimum capital
}
