% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_min_exposure}
\alias{f_min_exposure}
\title{Minimum Exposure}
\usage{
f_min_exposure(min_exposure, instr_risk, instrument_risk_target)
}
\arguments{
\item{min_exposure}{Notional exposure of the minimum trade}

\item{instr_risk}{Instrument risk as decimal fraction}

\item{instrument_risk_target}{Target risk as decimal fraction}
}
\description{
Minimum Exposure
}
