% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_normalization_factor}
\alias{f_normalization_factor}
\title{Calculate Normalization Factor}
\usage{
f_normalization_factor(raw_signal, target = 1)
}
\arguments{
\item{raw_signal}{A vector of signals on which the normalization is based.}

\item{target}{The target expected value of the signal scaled by the
normalization factor.}
}
\description{
The raw signal is scaled by a normalization factor.
f_normalization_factor() calculates the normalization factor based
on a vector of (previous) raw signals.
The normalization factor is

$$| \text{target} / E(\text{signal}) | $$

As a default the expected absolute value of the scaled values will be 1.
}
