% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_normalize_signal}
\alias{f_normalize_signal}
\title{Normalize Signal}
\usage{
f_normalize_signal(signal, normalization_factor = 1)
}
\arguments{
\item{normalization_factor}{}
}
\value{
Normalized signal as a number.
}
\description{
Normalize Signal
}
