% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_notional_exposure}
\alias{f_notional_exposure}
\title{Notional exposure}
\usage{
f_notional_exposure(
  combined_signal,
  capital,
  required_leverage_factor,
  instrument_weight
)
}
\arguments{
\item{combined_signal}{Combined signal. Should be normalized to a an expected
absolute value of 1 and clipped between -2 and 2.}

\item{capital}{Trading capital in account currency.}

\item{required_leverage_factor}{Required leverage factor. Indstrument risk
target divided by instrument risk (in same units).}

\item{instrument_weight}{Instrument weight.}
}
\value{
Number
}
\description{
Calculate the notional exposure.
}
