% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-portfolio.R
\name{f_portfolio_weighted_position}
\alias{f_portfolio_weighted_position}
\title{Calculate Portfolio Weighted Position}
\usage{
f_portfolio_weighted_position()
}
\value{
Number
}
\description{
Calculate weighted position.
}
