% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_position_change_ccy}
\alias{f_position_change_ccy}
\title{Calculate Position Change}
\usage{
f_position_change_ccy(position_table, position_size_ccy, t)
}
\arguments{
\item{position_table}{Position table}

\item{position_size_ccy}{Current (time \eqn{t}) position size in price units.}

\item{t}{Time index ("now")}
}
\value{
Amount in price units
}
\description{
Calculate position change in price units.
}
