% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-portfolio.R
\name{f_position_in_units}
\alias{f_position_in_units}
\title{Calculate Final Position To Trade In Units Of Contracts}
\usage{
f_position_in_units(target_position_size_units)
}
\arguments{
\item{target_position_size_units}{Target position to trade in units of
contracts.}
}
\value{
Number of contracts
}
\description{
Calculate final rounded position to trade in units of contracts. Truncates
the target position to the nearest integer towards 0.
}
