% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_price_returns}
\alias{f_price_returns}
\title{Calculate Price Returns}
\usage{
f_price_returns(prices, t)
}
\arguments{
\item{prices}{A price vector containing at least the two prices we need for
calculating returns. Oldest first. Top to bottom: Older to newer. The last
observation is time t.}

\item{t}{Time index ("now")}
}
\value{
A single number in price units.
}
\description{
Calculate price returns.
}
