% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_price_unit_vol}
\alias{f_price_unit_vol}
\title{Price Unit Volatility (instrument risk in price units)}
\usage{
f_price_unit_vol(price, inst_risk)
}
\arguments{
\item{inst_risk}{instrument risk.}

\item{prices}{A vector of prices in currency. Oldest first. Top to bottom:
Older to newer. The last observation is time t.}
}
\description{
Price Unit Volatility (instrument risk in price units)
}
