% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_price_volatility_ewma}
\alias{f_price_volatility_ewma}
\title{Instrument Risk In Units Of Percentage Of Price}
\usage{
f_price_volatility_ewma(current_price, last_ewma, window_length = 25)
}
\arguments{
\item{current_price}{Current price.}

\item{last_ewma}{Last exponential weighted moving average.}

\item{window_length}{Window length.}
}
\value{
Instrument risk in units of percentage of price
}
\description{
Price Volatility as Exponentially Weighted Moving Average (EWMA)
\itemize{
\item \eqn{E_{t}}: EWMA til tid \eqn{t}
\item \eqn{P_{t}}: Price at time \eqn{t}
\item \eqn{a}: Smoothing Parameter
\item \eqn{L}: Window length
\deqn{a = \frac{2}{1 + L}}
\deqn{E_{t} = a \cdot P_{t} + E_{t - 1}(1 - a)}
}

At f_price_volatility_ewma\eqn{t_0} use \code{last_ewma} = \code{current_price}.
}
