% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{f_profit_lvrg}
\alias{f_profit_lvrg}
\title{Growth With Leverage}
\usage{
f_profit_lvrg(V, L, r, b)
}
\arguments{
\item{V}{Value (\eqn{V_t}, value at time \eqn{t})}

\item{L}{Leverage factor.}

\item{r}{Return rate on investment.}

\item{b}{Borrowing rate.}
}
\value{
Profit in price currency.
}
\description{
Growth With Leverage
}
