% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{f_returns_from_prices}
\alias{f_returns_from_prices}
\title{Calculate net returns vector from price vector}
\usage{
f_returns_from_prices(prices)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}
}
\description{
Calculate net returns vector from price vector
}
