% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_risk_adjusted_signal}
\alias{f_risk_adjusted_signal}
\title{Instrument risk in units of annual standard deviation of price returns}
\usage{
f_risk_adjusted_signal(raw_signal, current_price, inst_risk)
}
\arguments{
\item{raw_signal}{Single number}

\item{current_price}{Price in currency as a single number}

\item{inst_risk}{Risk in units of annual standard deviation of price returns
as a single number}
}
\description{
\emph{Instrument risk} in units of \emph{annual standard deviation of price returns}
to match units of the rule.
}
