% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-statistics.R
\name{f_sd}
\alias{f_sd}
\title{Calculate Standard Deviation}
\usage{
f_sd(x, method = "unbiased", ...)
}
\arguments{
\item{x}{Vector.}

\item{method}{Calculation method.
\itemize{
\item \code{"unbiased"} Basic unbiased standard deviation.
\itemize{
\item @seealso \code{\link[stats:sd]{stats::sd()}}
}
\item \code{"ewa"} Based on Exponentially Weighted Average.
\itemize{
\item Additional parameters: \code{lambda}(number), \code{lookback} (integer).
\item @seealso \link{f_ewa_sd}
}
}}

\item{...}{}
}
\value{
Vector of moving standard deviations.
}
\description{
Calculate Standard Deviation
}
\examples{
f_sd(1:10, method = "ewa", lambda = 0.8, lookback = 3L)
}
