% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_sig_div_mul}
\alias{f_sig_div_mul}
\title{Signal Diversification Multiplier (SDM)}
\usage{
f_sig_div_mul(signal_correlations, signal_weights, min_cor = 0, max_sdm = 2.5)
}
\arguments{
\item{signal_correlations}{Correlations matrix for signals.}

\item{signal_weights}{Vector of signal weights.}

\item{min_cor}{Minimum value for each element in \code{signal_correlations}
matrix.}

\item{max_sdm}{Maximum output value of SDM.}
}
\value{
Scalar
}
\description{
\deqn{\text{SDM} = \frac{1}{\sqrt{W^{T}HW}}}
Where \eqn{H} is the correlation matrix for signals, and \eqn{W} is the
vector of instrument weights summing to one. We typically calculate the
SDM for each subsystem. E.g. for each instrument we calculate the SDM
using all the signals for that instrument instrument as input. The purpose
is to compensate for the reduction in volatility resulting from combining
multiple signals into one when we want to reach a particular risk target.

Recommended upper limit: 2.5. (Recommended by Robert Carver, not by the author
of rsystrade. See disclaimer in README.md.)

Negative correlations should be downwards clipped at 0 before calculation of
SDM.
}
