% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-alpha.R, R/formulas-risk.R
\name{f_stop_loss_gap}
\alias{f_stop_loss_gap}
\title{Calculate Stop Loss Gap}
\usage{
f_stop_loss_gap(price_unit_vol, stop_loss_fraction)

f_stop_loss_gap(price_unit_vol, stop_loss_fraction)
}
\arguments{
\item{stop_loss_fraction}{}
}
\description{
Calculate Stop Loss Gap

Stop loss gap
}
