% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-rules.R
\name{f_stop_loss_risky_capital_pct}
\alias{f_stop_loss_risky_capital_pct}
\title{Decimal fraction of capital at risk per trade for system with one instrument
only.}
\usage{
f_stop_loss_risky_capital_pct(risk_target, stop_loss_fraction)
}
\arguments{
\item{risk_target}{Annualized system risk target in percentage terms given
as a decimal fraction}

\item{stop_loss_fraction}{Stop loss fraction}
}
\description{
When using stop loss to exit trades using while trading one instrument only.
}
