% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_subsystem_pandl}
\alias{f_subsystem_pandl}
\title{Calculate Subsystem Profit And Loss}
\usage{
f_subsystem_pandl(position_table, instrument_return, t)
}
\arguments{
\item{position_table}{Position table.}

\item{t}{Time index ("now").}

\item{instrument_returns}{Instrument price returns. This is typically the
difference between today's and yesterday's price.}
}
\value{
A single number in price units.
}
\description{
Calculate subsystem profit and loss (subsystem returns).
}
