% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-portfolio.R
\name{f_target_position_size_units}
\alias{f_target_position_size_units}
\title{Calculate Target Position Size In Units Of Contracts}
\usage{
f_target_position_size_units(notional_exposure, price, fx_rate = 1)
}
\arguments{
\item{notional_exposure}{Notional exposure.}

\item{price}{Single price observation.}

\item{fx_rate}{FX rate.}
}
\value{
Number
}
\description{
Calculate the target position size in units of contracts.
}
