% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_total_cash}
\alias{f_total_cash}
\title{Calculate Cash Account Total}
\usage{
f_total_cash(position_tables, system_account_table, t)
}
\arguments{
\item{position_tables}{List of position tables.}

\item{t}{Time index ("now")}

\item{account_table}{Account table.}
}
\value{
Amount in price units
}
\description{
Calculate cash account total.
}
