% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_total_positions_changes_ccy}
\alias{f_total_positions_changes_ccy}
\title{Calculate Sum Of All Position Changes}
\usage{
f_total_positions_changes_ccy(position_tables, t)
}
\arguments{
\item{position_tables}{List of all position tables.}

\item{t}{Time index ("now")}
}
\value{
Amount in price units
}
\description{
Calculate sum of all position changes across all subsystems at time eqn{t}.
}
