% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_total_positions_value}
\alias{f_total_positions_value}
\title{Calculate Total Positions Value}
\usage{
f_total_positions_value(position_tables, t)
}
\arguments{
\item{position_tables}{List of all position tables.}

\item{t}{Time index ("now")}
}
\value{
Single number in units of price
}
\description{
Current aggregate value of all open positions.
}
