% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_total_returns}
\alias{f_total_returns}
\title{Calculate Total Price Returns}
\usage{
f_total_returns(position_tables, t)
}
\arguments{
\item{position_tables}{List of all position tables.}

\item{t}{Time index ("now")}
}
\value{
Aggregate return in units of price
}
\description{
Calculate total price returns across all instruments in system.
}
