% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_cor_mat_NAs}
\alias{fix_cor_mat_NAs}
\title{Fix NAs in correlation matrix}
\usage{
fix_cor_mat_NAs(cor_mat, value)
}
\arguments{
\item{cor_mat}{A correlation matrix}

\item{value}{The value to replace NA by}
}
\description{
Fix NAs in correlation matrix
}
