% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting.R
\name{get_all_position_sizes_ccy}
\alias{get_all_position_sizes_ccy}
\title{Get All Subsystem Position Sizes In Price Units From Position Tables}
\usage{
get_all_position_sizes_ccy(position_tables, t)
}
\arguments{
\item{position_tables}{Position tables}
}
\value{
Vector of position sizes in price units
}
\description{
Get all subsystem position sizes at time \eqn{t} in price units from position tables.
}
