% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting.R
\name{get_all_returns}
\alias{get_all_returns}
\title{Get Price Returns From All Instruments}
\usage{
get_all_returns(position_tables, t)
}
\arguments{
\item{position_tables}{List of all position tables.}

\item{t}{Time index ("now")}
}
\value{
Vector of price returns
}
\description{
Get price returns from all instruments at time \eqn{t}.
}
