% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting.R
\name{get_all_subsystem_pandls}
\alias{get_all_subsystem_pandls}
\title{Get All Subsystem Profits/Losses From Position Tables}
\usage{
get_all_subsystem_pandls(position_tables, t)
}
\arguments{
\item{position_tables}{List of all position tables.}
}
\value{
Vector of profits/losses in price units
}
\description{
Get all subsystem profits/losses in price units from position tables at time
\eqn{t}.
}
