% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_inst_names_by_parsed_algo}
\alias{get_inst_names_by_parsed_algo}
\title{Get Instrument Names By Parsed Algo}
\usage{
get_inst_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\description{
Get the instrument name from each parsed algo and place the names in a list
in the same order as they appear in the algos list.
}
