% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_modifier_list.R
\name{get_inst_names_by_position_modifier}
\alias{get_inst_names_by_position_modifier}
\title{Get Instrument Names From List Of Position Modifiers}
\usage{
get_inst_names_by_position_modifier(parsed_posmod_list)
}
\arguments{
\item{parsed_posmod_list}{}
}
\description{
Get Instrument Names From List Of Position Modifiers
}
