% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_multiplier_list.R
\name{get_inst_names_by_position_multiplier}
\alias{get_inst_names_by_position_multiplier}
\title{Get Instrument Names From Parsed List Of Position Multipliers}
\usage{
get_inst_names_by_position_multiplier(parsed_posmul_list)
}
\arguments{
\item{parsed_posmul_list}{}
}
\description{
Get Instrument Names From Parsed List Of Position Multipliers
}
