% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_multiplier_list.R
\name{get_multiplier_names_from_multiplier_list}
\alias{get_multiplier_names_from_multiplier_list}
\title{Get Multiplier Names From Parsed List Of Position Multipliers}
\usage{
get_multiplier_names_from_multiplier_list(parsed_posmul_list)
}
\arguments{
\item{parsed_posmul_list}{}
}
\description{
Get Multiplier Names From Parsed List Of Position Multipliers
}
