% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_num_rules_from_parsed_algos_list}
\alias{get_num_rules_from_parsed_algos_list}
\title{Get Number Of Unique Rules From Parsed Algos List}
\usage{
get_num_rules_from_parsed_algos_list(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\value{
Number
}
\description{
Get the number of unique rules from an expanded algos list.
}
