% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_num_rules_per_inst_from_parsed_algos}
\alias{get_num_rules_per_inst_from_parsed_algos}
\title{Get Number Of Rules Per Instrument From Parsed Algos}
\usage{
get_num_rules_per_inst_from_parsed_algos(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\value{
Named list of numbers of rules for each instrument.
}
\description{
Get a named list of numbers of rules for each instrument.
One named item for each instrument. The name of the item is the name of the
instrument.
Each item in the list is a number of rules for the corresponding instrument.
}
