% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{get_pos_mod_var_param_vals}
\alias{get_pos_mod_var_param_vals}
\title{Get Position Modifier Variable Parameter Values}
\usage{
get_pos_mod_var_param_vals(
  t,
  position_size_ccy,
  inst_data,
  position_modifier,
  pos_table_vars
)
}
\arguments{
\item{t}{Time index}

\item{position_size_ccy}{Position size in currency}

\item{inst_data}{Instrument data frame}

\item{position_modifier}{Position modifier}

\item{pos_table_vars}{Variables passed from update_position_table_row()}
}
\value{
Named list of param values
}
\description{
Get Position Modifier Variable Parameter Values
}
