% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{get_pos_mul_var_param_vals}
\alias{get_pos_mul_var_param_vals}
\title{Get Position Multiplier Variable Parameter Values}
\usage{
get_pos_mul_var_param_vals(
  t,
  position_size_ccy,
  inst_data,
  position_multipliers,
  pos_table_vars
)
}
\arguments{
\item{t}{Time index}

\item{position_size_ccy}{Position size in currency}

\item{inst_data}{Instrument data frame}

\item{pos_table_vars}{Variables passed from update_position_table_row()}

\item{position_multiplier}{Position multiplier from parsed multiplier list}
}
\value{
Nested list of param values for each multiplier function
}
\description{
Get Position Multiplier Variable Parameter Values
}
