% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_rule_function_names_by_parsed_algo}
\alias{get_rule_function_names_by_parsed_algo}
\title{Get Rule Function Names By Parsed Algo}
\usage{
get_rule_function_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{}
}
\value{
List
}
\description{
Get list of rule function names by parsed algo.
}
