% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_rule_names_by_parsed_algo}
\alias{get_rule_names_by_parsed_algo}
\title{Get Rule Names By Parsed Algo}
\usage{
get_rule_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\value{
List of character strings.
}
\description{
Gets the rule name as a character string from each parsed algo and place the
names in a list in the same order as they appear in the algos list. Each
rule is a combination of one signal rule and optionally one stop loss rule.

To get the names of all signal and stop loss rules, use
\code{get_rule_function_names_by_parsed_algo()}.

Rule names are the names of elements in the rule list for each algo.
}
