% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_rule_variation_names_by_parsed_algo}
\alias{get_rule_variation_names_by_parsed_algo}
\title{Get Rule Variation Names By Parsed Algo}
\usage{
get_rule_variation_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{}
}
\value{
List of rule variation names in the order they appear in the parsed
algos list
}
\description{
Get list of rule variation names by parsed algo.
}
