% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_signal_normalization_factors_by_algos}
\alias{get_signal_normalization_factors_by_algos}
\title{Get Signal Normalization Factors By Algos}
\usage{
get_signal_normalization_factors_by_algos(
  signal_normalization_factors,
  parsed_algos
)
}
\arguments{
\item{signal_normalization_factors}{Named list of unique normalization
factors.}
}
\value{
Unnamed list of normalization factors matching the order of the list
of parsed algos
}
\description{
Get signal normalization factors by algos.
}
