% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_unique_inst_paths_from_expanded_algos_list}
\alias{get_unique_inst_paths_from_expanded_algos_list}
\title{Get Unique Instrument Paths From Expanded Algos List}
\usage{
get_unique_inst_paths_from_expanded_algos_list(
  expanded_algos,
  instrument_data_folder_path
)
}
\arguments{
\item{expanded_algos}{Expanded list of algos.}

\item{instrument_data_folder_path}{Path to instrument data folder.}
}
\description{
Get the unique instrument paths from an expanded algos list.
}
