% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_unique_instr_names_from_parsed_algos_list}
\alias{get_unique_instr_names_from_parsed_algos_list}
\title{Get Unique Instrument Names From Expanded Algos List}
\usage{
get_unique_instr_names_from_parsed_algos_list(parsed_algos)
}
\arguments{
\item{parsed_algos}{}
}
\description{
Get Unique Instrument Names From Expanded Algos List
}
