% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_multiplier_list.R
\name{get_unique_multiplier_names_from_multiplier_list}
\alias{get_unique_multiplier_names_from_multiplier_list}
\title{Get Unique Multiplier Names From Parsed List Of Position Multipliers}
\usage{
get_unique_multiplier_names_from_multiplier_list(parsed_posmul_list)
}
\arguments{
\item{multiplier_list}{}
}
\description{
Get Unique Multiplier Names From Parsed List Of Position Multipliers
}
