% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_unique_rule_function_names_by_parsed_algo}
\alias{get_unique_rule_function_names_by_parsed_algo}
\title{Get Unique Rule Function Names By Parsed Algo}
\usage{
get_unique_rule_function_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos}
}
\value{
List
}
\description{
Get list of unique rule function names by parsed algo. These functions are
either signal rule functions or stop loss rule functions.
}
