% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_unique_rule_variation_names_by_parsed_algo}
\alias{get_unique_rule_variation_names_by_parsed_algo}
\title{Get Unique Rule Variation Names By Parsed Algo}
\usage{
get_unique_rule_variation_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos}
}
\value{
List
}
\description{
Get list of unique rule variation names by parsed algo.
}
