% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{get_variable_param_vals}
\alias{get_variable_param_vals}
\title{Get Values Of Variable Paramaters}
\usage{
get_variable_param_vals(t, inst_data, algo)
}
\arguments{
\item{inst_data}{Instrument data set as data frame}

\item{algo}{Algo}
}
\value{
Named list of parameter values
}
\description{
Get Values Of Variable Paramaters
}
