% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{load_instrument_data_sets}
\alias{load_instrument_data_sets}
\title{Load Instrument Data Sets}
\usage{
load_instrument_data_sets(parsed_algos, instrument_data_folder_path)
}
\arguments{
\item{parsed_algos}{Expanded list of algos.}

\item{instrument_data_folder_path}{Path to instrument data folder.}
}
\value{
List of data frames
}
\description{
Takes character string instrument names from each algo in a trade system
created by \code{make_system()} and loads the corresponding csv file into a list.
}
