% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{load_rule_functions}
\alias{load_rule_functions}
\title{Load Rule Functions}
\usage{
load_rule_functions(parsed_algos)
}
\arguments{
\item{parsed_algos}{}
}
\description{
DEPRECATED: This depends on a previous structure of parsed algos, where
rule functions were given by their name as a character string!

Load rule functions into a list. Functions are specified as functions
assigned to variables or by function names as character strings in parsed
algos. The functions must be loaded into the enclosing (typically global)
or package environment.
}
