% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{lwm}
\alias{lwm}
\title{Low water mark}
\usage{
lwm(prices, t, t_trade_open)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{t}{Time index.}

\item{t_trade_open}{Time index for the time when the trade is opened.}
}
\description{
Low water mark
}
