% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_multipliers.R
\name{m_block_same_direction_entry}
\alias{m_block_same_direction_entry}
\title{Block Entry In Same Direction}
\usage{
m_block_same_direction_entry(
  t,
  price,
  direction,
  t_last_position_entry,
  position_table,
  mode = 1
)
}
\arguments{
\item{t}{Time index.}

\item{price}{(Variable param) Price vector.}

\item{direction}{(Variable param) Direction. 1 for long, -1 for short, 0 for flat.}

\item{t_last_position_entry}{(Variable param) Time index of last position entry.}

\item{position_table}{(Variable param) Position table.}

\item{mode}{(Fixed param) Mode. See details.}
}
\description{
Position multiplier
}
\details{
\code{mode=1} is \code{block_same_dir_entry()}.

\code{mode=2} is \code{block_same_dir_entry_inside_watermarks()}.
}
