% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{mac_rule}
\alias{mac_rule}
\title{Calculate a moving average crossover signal.}
\usage{
mac_rule(
  prices,
  t = NA,
  ma_fast = NA,
  ma_slow = NA,
  n_fast = 20L,
  n_slow = 80L,
  gap = 0,
  strict = TRUE,
  binary = FALSE
)
}
\arguments{
\item{prices}{A vector of prices.}

\item{ma_fast}{A number. Fast \emph{moving average}. Vector or single numeric.}

\item{ma_slow}{A number. Slow \emph{moving average}. Vector or single numeric.}

\item{n_fast}{A positive integer. It is the responsibility of the user to
check that the input value makes sense.}

\item{n_slow}{A positive integer. It is the responsibility of the user to
check that the input value makes sense.}

\item{gap}{A positive integer. Gap size in same unit as the parameters above
(typically days).}

\item{strict}{Boolean. If \code{strict=TRUE}, \code{n_slow} must be smaller than the
number of prices in the \code{prices} vector, and \code{n_slow} must be greater than
\code{n_fast}.}

\item{binary}{If \code{TRUE}: Binary mode. If \code{FALSE}: Proportional signal.
In binary mode returns
\itemize{
\item 1 when ma_fast > ma_slow, and abs(ma_fast - ma_slow) > gap.
\item -1 when ma_fast < ma_slow, and abs(ma_fast - ma_slow) > gap.
\item 0 when abs(ma_fast - ma_slow) < gap.
}

1 indicates uptrend i.e. go long. -1 indicates downtrend i.e. go short.}
}
\value{
A trade signal
}
\description{
If \code{ma_fast} or \code{ma_slow} are not provided, both values are calculated from
price vector. In this case \code{n_fast} and \code{n_slow} \emph{must} be provided.

If both \code{ma_fast} and \code{ma_slow} are provided, \code{n_slow} and \code{n_slow} will be
ignored.

If \code{strict=FALSE} the function will not fail if \code{n_fast > n_slow}, but a
warning will be given. Also, if \code{strict=FALSE} and the length of the price
vector is smaller than \code{n_slow}, \code{n_slow} will be set to the length of
the price vector.

\code{ma_fast} and \code{ma_slow} inputs are typically single steps of moving averages,
but can be vectors.

On a side note, \code{mac_rule()} doesn't check that \code{ma_fast} and \code{ma_slow}
inputs are in fact calculated as a moving averages. Any number or vector
will work.
}
