% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_list_names_unique}
\alias{make_list_names_unique}
\title{Make List Names Unique}
\usage{
make_list_names_unique(list)
}
\arguments{
\item{list}{Named list}
}
\value{
Named list with unique names
}
\description{
Appends \code{.#} to duplicate list names, where \verb{#} represents a consequtive
number.
}
