% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_modifier_list.R
\name{make_position_modifiers_list}
\alias{make_position_modifiers_list}
\title{Make List Of Position Modifiers From Parsed Position Modifiers List}
\usage{
make_position_modifiers_list(position_modifiers = list(), inst_names)
}
\arguments{
\item{position_modifiers}{User provided named list of position modifier
functions.}

\item{inst_names}{Vector of instrument names.}
}
\value{
Named list
}
\description{
Creates a named list. Values are all \code{NA} and the names are the names of
instruments (or any vector of character strings provided as input).
}
