% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_multiplier_list.R
\name{make_position_multipliers_list}
\alias{make_position_multipliers_list}
\title{Make List Of Position Multipliers From Position Multipliers List}
\usage{
make_position_multipliers_list(position_multipliers = list(), inst_names)
}
\arguments{
\item{position_multipliers}{User provided named list of position multiplier
functions.}

\item{inst_names}{Vector of instrument names.}
}
\value{
Named list
}
\description{
Creates a named list. Values are all \code{NA} and the names are the names of
instruments (or any vector of character strings provided as input).
}
\details{
Initialize list of position multipliers User may provide a named list where
the names are instrument names and values are position multiplier functions.

\code{make_position_multifiers_list()} creates a named list. Each element corresponds
to an instrument, in the order instruments occur in the inst_data list. The
name of each element is the name of the instrument. User provided position
multiplier functions will be assigned to the appropriate element in the
list.

Any position multiplier function for which the name doesn't match any
instrument in the system, will be ignored.

If no position multiplier function is assigned to an instrument, the position
of that instrument will not be multiplied.

Position multiplier functions must comply with these specifications:
\itemize{
\item The first parameter must be \code{t}, current time index. \code{t} is provided by
\code{modify_position()}.
\item The second parameter must be \code{position_size_ccy}. \code{position_size_ccy} is
provided by \code{modify_position()}. \code{position_size_ccy} may be a modified
position calculated by a position modifier function in \code{modify_position()},
not necessarily the raw  \code{position_size_ccy} input to \code{modify_position()}.
\item The first two parameters are variable parameters. This means that they are
provided with data as the system is being updated. Parameters that are not
fixed are variable. I.e. any parameters that are not provided with a value
when running \code{make_system()}, is a variable parameter.
\item Any parameters that are provided with a value when running \code{make_system()},
is a fixed parameter.
}

First element in each input \code{position_multipliers[[i]]$multiplier[[m]]} list
is the multiplier name, second element is the multiplier function
definition itself. Remaining elements are the fixed params.

Note: You may provide any name as multiplier name. If the multiplier
function is assigned to a variable, you may put this variable as the
function definition. So the function name doesn't have to match the
multiplier name. The multiplier name is just a label for your convenience.

Note: Variable params are not provided by user, i.e. do not appear
in multiplier list.

If the list of data sets only contains \code{"inst1"} and \code{"inst2"}, then \code{"inst3"}
will be ignored when parsing the \code{position_multipliers} list.

The input format of the \code{position_multipliers} list provided to
\code{make_system()} is (example):

\if{html}{\out{<div class="sourceCode R">}}\preformatted{position_multipliers <- list(
  list(
    instruments = list("inst1", "inst2"),
    multipliers = list(
     list(
       "f1",
        f1,
        y1 = 10
     ),
     list(
       "f2",
        f2,
        y2 = 20
     )
    )
  ),
  list(
    instruments = list("inst3", "inst4"),
    multipliers = list(
     list(
       "f3",
        f3,
        y3 = 30
     ),
     list(
       "f4",
        f4,
        y4 = 40
     )
    )
  )
)
}\if{html}{\out{</div>}}

This will be parsed into a list formatted as

\if{html}{\out{<div class="sourceCode R">}}\preformatted{position_multipliers <- list(
  inst1 = list(
   list(
     multiplier_name = "f1",
     multiplier_function = f1,
     variable_params = list(
        x1 = "x1"
     ),
     fixed_params = list(
        y1 = 10
     )
   ),
   list(
     multiplier_name = "f2",
     multiplier_function = f2,
     variable_params = list(
        x2 = "x2"
     ),
     fixed_params = list(
        y2 = 20
     )
   )
  ),
  inst2 = list(
   list(
     multiplier_name = "f1",
     multiplier_function = f1,
     variable_params = list(
        x1 = "x1"
     ),
     fixed_params = list(
        y1 = 10
     )
   ),
   list(
     multiplier_name = "f2",
     multiplier_function = f2,
     variable_params = list(
        x2 = "x2"
     ),
     fixed_params = list(
        y2 = 20
     )
   )
  ),
  inst3 = list(
   list(
     multiplier_name = "f3",
     multiplier_function = f3,
     variable_params = list(
        x3 = "x3"
     ),
     fixed_params = list(
        y3 = 30
     )
   ),
   list(
     multiplier_name = "f4",
     multiplier_function = f4,
     variable_params = list(
        x4 = "x4"
     ),
     fixed_params = list(
        y4 = 40
     )
   )
  ),
  inst4 = list(
   list(
     multiplier_name = "f3",
     multiplier_function = f3,
     variable_params = list(
        x3 = "x3"
     ),
     fixed_params = list(
        y3 = 30
     )
   ),
   list(
     multiplier_name = "f4",
     multiplierr_function = f4,
     variable_params = list(
        x4 = "x4"
     ),
     fixed_params = list(
        y4 = 40
     )
   )
  )
)
}\if{html}{\out{</div>}}
}
