% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{min_exposure}
\alias{min_exposure}
\title{Minimum exposure}
\usage{
min_exposure(min_exposure, instr_risk, target_risk)
}
\arguments{
\item{min_exposure}{Notional exposure of the minimum trade}

\item{instr_risk}{Instrument risk as decimal fraction}

\item{target_risk}{Target risk as decimal fraction}
}
\description{
Minimum exposure
}
