% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-alpha.R
\name{moving_average}
\alias{moving_average}
\title{Moving Average}
\usage{
moving_average(prices, n, t = NA)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{n}{Window length.}

\item{t}{Time index.}
}
\value{
A number.
}
\description{
Calculates average of n items prior to time t in price vector.
N (length of price vector) smaller than or equal to n is accepted.
This will not be the desired moving average, but will also not fail.
}
