% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-alpha.R
\name{moving_average_crossover}
\alias{moving_average_crossover}
\title{Moving Average Crossover}
\usage{
moving_average_crossover(ma_fast, ma_slow, gap = 0)
}
\arguments{
\item{ma_fast}{Fast moving average}

\item{ma_slow}{Slow moving average}

\item{gap}{Gap.}
}
\value{
A boolean.
}
\description{
Returns TRUE when ma_fast > ma_slow.
1 indicates uptrend ie. go long.
-1 indicates downtrend ie. go short.
}
