% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{multiply_position}
\alias{multiply_position}
\title{Multiply Position By Multiplier Functions}
\usage{
multiply_position(mul_variable_param_vals, position_multipliers)
}
\arguments{
\item{mul_variable_param_vals}{Multiplier variable param values.}

\item{position_multipliers}{A list of multipliers affecting a single
instrument position.}
}
\value{
List of multiplier function outputs. The first element must be the
multiplier value.
}
\description{
Apply multipliers from parsed list of position multipliers to a single
position.
}
\details{
Note: If different multiplier functions are applied to the same instrument,
and these functions have elements in their outputs with identical names,
the column names in \code{position_table} will be modified. E.g.
if function \code{f1} outputs \code{list(combined_multiplier = 2, x = 10, y = 20)}
and function \code{f2} outputs \code{list(combined_multiplier = 3, x = 10, z = 20)},
then this will result in the following columns in \code{position_table}:
\code{combined_multiplier}, \code{x}, \code{y}, \code{x.1} and \code{z}.
}
