% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{notional_exposure}
\alias{notional_exposure}
\title{Notional exposure
F14}
\usage{
notional_exposure(risk_target, capital, instrument_risk)
}
\arguments{
\item{risk_target}{Risk target as decimal fraction}

\item{capital}{Trading capital in currency}

\item{instrument_risk}{Instrument risk as decimal fraction}
}
\description{
Notional exposure
F14
}
