% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_multipliers.R
\name{o_limit_cor_risk}
\alias{o_limit_cor_risk}
\title{Limit Shock Portfolio Correlation Risk}
\usage{
o_limit_cor_risk(t, position_tables, max_risk = NA, capital)
}
\arguments{
\item{t}{Time index.}

\item{position_tables}{Position tables list.}

\item{max_risk}{Maximum risk. Risk limit.}

\item{capital}{Total system capital.}
}
\value{
List of multiplier output. The first element is the multiplier
value in \eqn{[0, 1]}. The second element is \code{cor_shock_pf_risk}.
}
\description{
Calculate a limit multiplier for shock portfolio correlation. The assumption
here is that if absolute correlation gets close to 1, the portfolio risk
will approximately be be the sum of weighted instrument risks. Further, as
the absolute correlation decreases, so will the sum of weighted instrument
risks. The smaller the absolute correlation, the bigger the difference
between the portfolio risk and the sum of weighted instrument risks.
}
\details{
The idea is to limit the sum of weighted standard risks to a maximum value.

A way to determine the \code{max_risk} could be to calculate the (worst) 99th
percentile of \emph{correlation shock portfolio risk}.
}
