% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_multipliers.R
\name{o_limit_leverage_risk}
\alias{o_limit_leverage_risk}
\title{Limit Portfolio Leverage}
\usage{
o_limit_leverage_risk(t, position_tables, max_leverage = NA, capital)
}
\arguments{
\item{t}{Time index.}

\item{position_tables}{Position tables list.}

\item{max_leverage}{Maximum leverage. Upper leverage limit.}

\item{capital}{Total system capital.}
}
\value{
List of multiplier output. The first element is the multiplier
value in \eqn{[0, 1]}. The second element is \code{pf_leverage}.
}
\description{
Limit Portfolio Leverage
}
\details{
The idea is to limit the portfolio leverage to a maximum value.

Portfolio leverage is calculated as the total value of all positions in the
currency of the account divided by the system capital.

A way to determine the \code{max_leverage} could be to calculate the (worst) 99th
percentile of \emph{portfolio leverage} over time.
}
