% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_modifiers.R
\name{p_stop_loss}
\alias{p_stop_loss}
\title{Stop Loss Position Modifier}
\usage{
p_stop_loss(
  t,
  position_size_ccy,
  price,
  instrument_risk,
  t_last_position_entry,
  direction,
  stop_loss_fraction = 0.5,
  rnd = FALSE
)
}
\arguments{
\item{t}{Time index.}

\item{position_size_ccy}{Position size in account currenct.}

\item{price}{(Variable param) A vector of prices in currency. Oldest first. Top to bottom:
Older to newer.}

\item{instrument_risk}{(Variable param) Instrument risk at time t-1.}

\item{t_last_position_entry}{(Variable param) Time index of the time when current trade was entered.}

\item{direction}{(Variable param) Was current trade long or short at time t-1? 1 for long, -1
for short.}

\item{stop_loss_fraction}{(Fixed param) Stop loss fraction.}

\item{rnd}{(Fixed param) If TRUE, add small random amount to stop loss level. Negative if
short.}
}
\value{
A named list containing:
\itemize{
\item \code{signal} Trade on/off indicator: 1 or 0
\item \code{stop_loss_gap} Stop loss gap
}
}
\description{
Return 0 if price level breached stop loss threshold. Return input
position_size_ccy if stop loss threshold was not breached.
}
