% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_portfolio_multiplier_list.R
\name{parse_portfolio_multipliers_list}
\alias{parse_portfolio_multipliers_list}
\title{Parse List Of Portfolio Multipliers}
\usage{
parse_portfolio_multipliers_list(portfolio_multipliers)
}
\arguments{
\item{portfolio_multipliers}{Input list of portfolio multipliers.}
}
\value{
Output list of portfolio multipliers
}
\description{
Parse a list of portfolio multiplier specification lists.

Portfolio multiplier specification lists are provided by the user as
parameter \code{portfolio_multipliers} to \code{make_system()}.
}
\details{
Each input portfolio multiplier specification list specifies a multiplier
name, a multiplier function and any number of fixed parameters. The parser
automatically determines which parameters of each multiplier function are
fixed parameters provided by the user, and which are variable parameters
provided by the system.

Input format:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{portfolio_multipliers = list(
  list(
    <multiplier-name-1>,
    <multiplier-function-1>,
    <fixed-params-1_1>,
    <fixed-params-1_2>
  ),
  list(
    <multiplier-name-2>,
    <multiplier-function-2>,
    <fixed-params-2_1>,
    <fixed-params-2_2>
  )
)
}\if{html}{\out{</div>}}

Output format:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{portfolio_multipliers = list(
  list(
    modifier_name = <multiplier-name-1>,
    modifier_function = <multiplier-function-1>,
    variable_params = list(
      <variable-param-1_1>,
      <variable-param-1_2>
    )
    fixed_params = list(
      <fixed-params-1_1>,
      <fixed-params-1_2>
     )
  ),
  list(
    modifier_name = <multiplier-name-2>,
    modifier_function = <multiplier-function-2>,
    variable_params = list(
      <variable-param-2_1>,
      <variable-param-2_2>
    )
    fixed_params = list(
      <fixed-params-2_1>,
      <fixed-params-2_2>
     )
  )
  )
)
}\if{html}{\out{</div>}}

Unlike \emph{position multipliers} and \emph{position modifiers}, variable params for
\emph{portfolio multipliers} don't take input directly from \code{inst_data}.
Instrument data can be accessed via the system variables, which are
available to portfolio multiplier functions.

System variables which are available to portfolio multiplier functions are:
\itemize{
\item \code{trade_system} (everything in the system up until t - 1)
\item \code{t}
\item \code{signal_weights_all_algos}
\item \code{sig_norm_fact_by_algos}
\item \code{signal_tables}
\item \code{subsystem_pandl_vectors}
\item \code{instrument_list}
\item \code{subsystem_ret_cor_win_len}
\item \code{subsystem_ret_cor_mat}
\item \code{instrument_weights}
\item \code{combined_signals_and_sdm}
\item \code{raw_combined_signals}
\item \code{signal_div_mul_vect}
\item \code{inst_div_mul}
\item \code{instrument_risk_target}
\item \code{position_tables}
}
}
