% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{price_unit_vol}
\alias{price_unit_vol}
\title{Price unit volatility (instrument risk in price units)}
\usage{
price_unit_vol(instr_risk, price)
}
\arguments{
\item{instr_risk}{Instrument risk.}

\item{price}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}
}
\description{
Price unit volatility (instrument risk in price units)
}
