% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{prices_from_returns}
\alias{prices_from_returns}
\title{Generate price vector from returns vector}
\usage{
prices_from_returns(returns, initial_price)
}
\arguments{
\item{returns}{Returns.}

\item{initial_price}{Initial price.}
}
\value{
Vector of prices in currency of initial price.
}
\description{
Generate price vector from returns vector
}
