% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{profit_lvrg}
\alias{profit_lvrg}
\title{Growth with leverage}
\usage{
profit_lvrg(V, L, r, b)
}
\arguments{
\item{V}{Value (V_t, value at time t)}

\item{L}{Leverage factor.}

\item{r}{Return rate on investment.}

\item{b}{Borrowing rate.}
}
\value{
Profit in price currency.
}
\description{
Growth with leverage
}
