% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{returns_from_prices}
\alias{returns_from_prices}
\title{Generate net returns vector from price vector}
\usage{
returns_from_prices(prices)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}
}
\description{
Generate net returns vector from price vector
}
