% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{risky_capital_pct}
\alias{risky_capital_pct}
\title{Decimal fraction of capital at risk per trade for Starter System
F15, LT, p. 113}
\usage{
risky_capital_pct(risk_target, stop_loss_fraction)
}
\arguments{
\item{risk_target}{Risk target as decimal fraction}

\item{stop_loss_fraction}{Stop loss fraction}
}
\description{
Decimal fraction of capital at risk per trade for Starter System
F15, LT, p. 113
}
