% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rolling_window}
\alias{rolling_window}
\title{Apply Function To A Rolling Window}
\usage{
rolling_window(x, first_t = 2, last_t = NA, window_length, func, mode = 1, ...)
}
\arguments{
\item{x}{A vector, e.g. a column in a data frame. Newest first. Top to
bottom: Newer to older.}

\item{first_t}{Time index of the first observation in the window.}

\item{last_t}{Time index of the last observation in the window.}

\item{window_length}{Moving Average window length.}

\item{func}{Function to be applied to window.}

\item{mode}{Rolling mode
\itemize{
\item \code{1} Time index range of window is \eqn{[t -} window_length\eqn{, t - 1]}.
\item \code{2} Time index range of window is \eqn{[t -} window_length \eqn{+ 1, t]}.
}}

\item{...}{}
}
\value{
A vector of length \code{length(x)}  if \code{mode==1} and \code{length(x) + 1} if
\code{mode==2}.
}
\description{
Applies a function to a rolling window of a vector as the time index \code{t}
moves through the vector. For instance, if the input \code{func} is \code{f_average},
then \code{rolling_window()} will calculate a moving average.

If the window length is shorter than the length of the vector up until \code{t}
for a step, the window length will be set to the length of the vector, and
a warning will be given. The function will not fail or abort in this case.
}
\details{
The first element of the output vector is always \code{NA}, since there will be no
elements in the window at time \eqn{t = 1}.
}
