% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rolling_window_step}
\alias{rolling_window_step}
\title{Apply Function To A Rolling Window Step}
\usage{
rolling_window_step(t = NA, x, last_win_id = NA, window_length, func, ...)
}
\arguments{
\item{t}{Time index.}

\item{x}{A vector. Oldest first. Top to bottom: Older to newer.}

\item{last_win_id}{Time index of the last observation in the window.}

\item{window_length}{Window length.}
}
\value{
A single number (or whatever is the output of \code{func})
}
\description{
Applies a function to a time window of a vector. For instance, if the input
\code{func} is \code{f_average}, then \code{rolling_window()} will calculate a single
step of a moving average, i.e. the mean of the data in the window.

If the window length is shorter than the length of the vector, the window
length will be set to the length of the vector, and a warning
will be given. The function will not fail or abort in this case.
}
\details{
If we are at time \code{t}, and we want to calculate the average for the interval
\eqn{[t - \text{window length}, t - 1]}, we can set \code{t} to the time index
immediately subsequent to the last observation in the window and let
\code{last_x_id} be \code{NA}.

A typical use case, however, is to calculate a moving average of the window
\eqn{[t - L + 1, t]}, where \eqn{L} is the length of the window. In this
case we could let \code{t} be \code{NA} and set \code{last_x_id} to the current time index.

If \code{t} is given a value and \code{last_x_id} is \code{NA}, the last observation in
the window will be \code{t - 1}. Alternatively one could set \code{last_x_id} and
let \code{t} be \code{NA}. If \code{t} and \code{last_x_id} are both \code{NA}, \code{last_x_id}
will be set to the length of the data vector. If \code{t} and \code{last_x_id} are
both given a value, \code{t} will be ignored.
}
