% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsystrade-package.R
\docType{package}
\name{rsystrade-package}
\alias{rsystrade}
\alias{rsystrade-package}
\title{rsystrade: Framework For Testing Systematic Strategies And Automated Trading}
\description{
Framework for researching, testing and executing systematic trading strategies. Get live data from broker API or use simulated or other data. Apply provided or bespoke rules. Perform statistical analyses. Let the system trade automatically. Get plots and reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mahovo/rsystrade}
  \item Report bugs at \url{https://github.com/mahovo/rsystrade/issues}
}

}
\author{
\strong{Maintainer}: Martin Hoshi Vognsen \email{martinvognsen@gmail.com} (\href{https://orcid.org/0000-0003-4460-7763}{ORCID})

}
\keyword{internal}
