% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{run_system}
\alias{run_system}
\title{Run System}
\usage{
run_system(system, mode = "sim")
}
\arguments{
\item{system}{List containing a full system.}

\item{mode}{"live" for live trading or "sim" for simulation mode.}
}
\value{
List containing full updated system.
}
\description{
Run System
}
