% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_generators.R
\name{s_mac_stoploss}
\alias{s_mac_stoploss}
\title{Moving Average Crossover Stop Loss}
\usage{
s_mac_stoploss(
  t = t,
  price = price,
  signal_table,
  position_table,
  config,
  ma_fast = NA,
  ma_slow = NA,
  n_fast = 25L,
  n_slow = 100L,
  gap = 0,
  strict = TRUE,
  binary = FALSE
)
}
\arguments{
\item{t}{Time index after the latest price in the windows. (Note that
the latest price is the same for both fast and slow window.)}

\item{price}{A vector of prices in currency. Oldest first. Top to bottom:
Older to newer. The last observation is time t.}

\item{signal_table}{Signal table}

\item{position_table}{Position table}

\item{config}{Config list}

\item{ma_fast}{A number. Fast \emph{moving average}. Vector or single numeric.}

\item{ma_slow}{A number. Slow \emph{moving average}. Vector or single numeric.}

\item{n_fast}{A positive integer. It is the responsibility of the user to
check that the input value makes sense.}

\item{n_slow}{A positive integer. It is the responsibility of the user to
check that the input value makes sense.}

\item{gap}{A positive integer. Gap size in same unit as the parameters above
(typically days).}

\item{strict}{Boolean. If \code{strict=TRUE}, \code{n_slow} must be smaller than the
number of prices in the \code{prices} vector, and \code{n_slow} must be greater than
\code{n_fast}.}

\item{binary}{If \code{TRUE}: Binary mode. If \code{FALSE}: Proportional signal.
In binary mode returns
\itemize{
\item 1 when ma_fast > ma_slow, and abs(ma_fast - ma_slow) > gap.
\item -1 when ma_fast < ma_slow, and abs(ma_fast - ma_slow) > gap.
\item 0 when abs(ma_fast - ma_slow) < gap.
}}
}
\description{
Moving Average Crossover Stop Loss
}
