% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{trade_sim}
\alias{trade_sim}
\title{Execute trades for simulation}
\usage{
trade_sim(
  prices,
  init_capital = 1e+05,
  n_fast = 16,
  n_slow = 64,
  risk_target = 0.12,
  risk_window_length = 25,
  stop_loss_fraction = 0.5
)
}
\arguments{
\item{n_fast}{n for fast moving average}

\item{n_slow}{n for slow moving average}

\item{stop_loss_fraction}{Generates a dataframe of simulated trades from time series of prices.}
}
\description{
Execute trades for simulation
}
