% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_position_table}
\alias{update_position_table}
\title{Update Position Table}
\usage{
update_position_table(signal_table, position_table)
}
\arguments{
\item{...}{}
}
\description{
Based on signal:
\itemize{
\item If a trade is not on: Determine whether to enter a trade, and if so
in which direction.
\item If a trade is on: Determine whether to exit the trade or to switch
direction.
\item It is not allowed to enter a trade in the same direction as the
previous trade.
}

Update the position table accordingly.
}
