% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_position_table_row}
\alias{update_position_table_row}
\title{Update Position Table Row}
\usage{
update_position_table_row(
  t,
  inst_data,
  position_table,
  position_modifier,
  position_multipliers,
  raw_combined_signal,
  signal_div_mul,
  instrument_weight,
  inst_div_mul,
  instrument_risk_target,
  subsystem_ret_cor_mat,
  system_account_table,
  config
)
}
\arguments{
\item{t}{Time index.}

\item{inst_data}{Instrument data table.}

\item{position_table}{Position table.}

\item{position_modifier}{Position modifier function.}

\item{raw_combined_signal}{Raw combined signal value.}

\item{signal_div_mul}{Signal diversification multiplier.}

\item{subsystem_ret_cor_mat}{Subsystem returns correlation matrix.}

\item{system_account_table}{System account table.}

\item{config}{Configuration list.}

\item{signal_table}{Signal table.}
}
\value{
List
}
\description{
Based on signal:
\itemize{
\item If a trade is not on: Determine whether to enter a trade, and if so
in which direction.
\item If a trade is on: Determine whether to modify the position size, exit the
trade or to switch direction.
}

Update the position table accordingly.
}
