% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_signal_normalization_factors}
\alias{update_signal_normalization_factors}
\title{Update Signal Normalization Factors}
\usage{
update_signal_normalization_factors(
  parsed_algos,
  signal_tables,
  instrument_data_sets,
  target = 1,
  method = "equal",
  ...
)
}
\arguments{
\item{parsed_algos}{Parsed algos list from trade system.}

\item{signal_tables}{Signal tables list from trade system.}

\item{instrument_data_sets}{Instrument data sets list from trade system.}

\item{target}{The target expected value of the signal scaled by the
normalization factor. Default is 1.}

\item{method}{Method.
\itemize{
\item \code{"equal"} All rules use the same normalization factor, passed as
\code{target}.
\item \code{"pool_traded"} Calculate the normalization factor for each rule based on
actual past signals for that rule across all the instruments to which
the individual rule has been applied.
We are pooling all the instruments for each rule. We are not taking the
cross section median across instruments, as we do for \code{pool_all()},
because the number of instruments per rule is likely to be small - taking
the median of two values doesn't make much sense.
\item \code{"pool_all"} Calculate the normalization factor for each rule based on
signals simulated by applying each rule to all available instruments and
pooling the resulting signals.
The normalization factor for each rule is the target divided by the mean
absolute value of the pooled signals.
Additional argument:
\itemize{
\item \code{min_period}. A value of \code{250} (ca. 1 year of
daily data) might be a good starting point. It is up to the user to
provide data sets with enough data.
}
\item \code{"median_pool_all"} Calculate the normalization factor for each rule
based on signals simulated by applying each rule to all available
instruments.
The normalization factor for each rule is the target divided by the mean
absolute value of all cross section medians.
Additional argument:
\itemize{
\item \code{min_period}. A value of \code{250} (ca. 1 year of
daily data) might be a good starting point. It is up to the user to
provide data sets with enough data.
}
\item \code{"pool_class"} Calculate the normalization factor for each rule based on
signals simulated by applying each rule to all available instruments in
a relevant asset class.
}}

\item{...}{Additional method specific arguments.}
}
\value{
Named list of normalization factors
}
\description{
Each raw signal is scaled by a normalization factor. This normalization
factor is the \emph{required leverage target}.

\code{update_signal_normalization_factors()} calculates the normalization factors
and updates the \code{signal_normalization_factors} list.
}
