% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_system}
\alias{update_system}
\title{Update System}
\usage{
update_system(trade_system, t, ...)
}
\arguments{
\item{t}{Time index.}

\item{...}{Additional arguments.}

\item{system}{Trade system list containing a complete system.}
}
\description{
Config options
\code{signal_weight_calculation_method}
\itemize{
\item \code{"equal"} (default)
\item \code{"estimate"}
}
}
