% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_time}
\alias{update_time}
\title{Update time}
\usage{
update_time(t, inst_data, algo)
}
\arguments{
\item{t}{Time index.}

\item{inst_data}{List of unique data frames.}

\item{algo}{Algo list for a single expanded algo.}
}
\description{
Get latest time from data frame in \code{algo} list.
}
