% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{apply_portfolio_multiplier}
\alias{apply_portfolio_multiplier}
\title{Apply Portfolio Multiplier To Positions}
\usage{
apply_portfolio_multiplier(
  t,
  prices,
  modified_target_pos_ccy,
  portfolio_multiplier_value,
  t_last_position_entry,
  latest_trade_direction,
  position_tables,
  config
)
}
\arguments{
\item{t}{Positive integer. Time index.}

\item{prices}{A vector of prices in currency. Oldest first. Top to bottom:
Older to newer. The last observation is time t.}

\item{modified_target_pos_ccy}{A number. Modified target position in currency
amount.}

\item{portfolio_multiplier_value}{Portfolio multiplier value.}

\item{t_last_position_entry}{Positive integer. Time index of last previous
position entry.}

\item{latest_trade_direction}{-1, 0 or 1. Latest trade direction.}

\item{position_tables}{Position tables}

\item{config}{Config list}
}
\value{
List of updated position tables
}
\description{
Multiply position sizes by portfolio multiplier and update position tables.
}
