% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{backfill_table}
\alias{backfill_table}
\title{Backfill Table With NAs}
\usage{
backfill_table(table, new_row)
}
\arguments{
\item{new_row}{New row containing elements beyond the columns in the signal
table}

\item{signal_table}{Signal table}
}
\value{
A named list
}
\description{
When a list of values is returned by a signal generator or a position
modifier, and the list contains more elements than there are columns in
the corresponding signal/position table, new columns are appended to the
table and previous rows are backfilled with NAs in those columns.

This situation occurs when a signal generator outputs values in addition to
the trade signal, or when a position modifier outputs values in addition to
the modified position.
}
