% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{close_trade_stop_loss}
\alias{close_trade_stop_loss}
\title{Open trade for Moving Average Crossover}
\usage{
close_trade_stop_loss(
  prices,
  t,
  t_trade_open,
  stop_loss_gap,
  direction = 0,
  rnd = FALSE
)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{t}{Time index.}

\item{t_trade_open}{Time index of the time when trade is opened.}

\item{stop_loss_gap}{Stop loss gap.}

\item{direction}{Is current trade long or short? 1 for long, -1 for short.}

\item{rnd}{If TRUE, add small random amount to stop loss level. Negative if
short.}

\item{moving_average_crossover}{1 for long, 0 for short.}

\item{latest_trade_direction}{1 for long, 0 for short.}
}
\value{
-1, 0 or 1.
}
\description{
Returns 1 for going long and -1 for going short.
Returns 0 for no change.
}
\details{
1 if latest_trade_direction = -1 and moving_average_crossover = TRUE.
1 if latest_trade_direction = 0 and moving_average_crossover = TRUE.
-1 if latest_trade_direction = 1 and moving_average_crossover = FALSE.
-1 if latest_trade_direction = 0 and moving_average_crossover = FALSE.
0 if latest_trade_direction = 1 and moving_average_crossover = TRUE.
0 if latest_trade_direction = -1 and moving_average_crossover = FALSE.

No trade will ever open when direction = 0, so latest_trade_direction can only
ever be 0 before the first trade.

moving_average_crossover:
TRUE indicates uptrend ie. go long.
FALSE indicates downtrend ie. go short.
}
