% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{combine_signals}
\alias{combine_signals}
\title{Combine Signals For Instrument Subsystem}
\usage{
combine_signals(signal_tables, algos, min_cor, max_sdm, t)
}
\arguments{
\item{signal_tables}{List of data frames. One data frame for each signal.}

\item{algos}{List of parsed (expanded) algos.}

\item{min_cor}{Minimum correlation.}

\item{max_sdm}{Maximum signal diversification multiplier.}

\item{t}{Time index.}
}
\value{
List containing a vector of combined signals and a vector of signal
diversification multiplier values
}
\description{
For each instrument subsystem calculates a weighted sum of all signals in
that subsystem.

A subsystem is one or more rules applied to one instrument. A signal is the
output from a rule.

\code{combine_signals()} returns a single combined signal for each instrument.
Now the subsystem consists of exactly one rule applied to one instrument.
}
