% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{expand_algos}
\alias{expand_algos}
\title{Expand algos}
\usage{
expand_algos(algos)
}
\arguments{
\item{algos}{A nested list (see specification above)}
}
\value{
A list where each element is a list containing one element named
\code{instrument} and one list named \code{rule} containing one signal rule and one
optional stop loss rule.
}
\description{
Takes a nested list of algo specifications and expands it to a list of all
possible permutations.

An algo is a pair consisting of an instrument and a rule.

A "rule" is a combination of one signal rule and optionally one stop loss
rule.

The \code{algos} parameter is a nested list in one in the following forms.
Pairs of one instrument and one rule:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{algos <- list(
  list( ## algo 1: algos[[1]]
    instruments = list("instr1"),
    rules = list(rule1)
  ),
  list( ## algo 2: algos[[2]]
    instruments = list("instr2"),
    rules = list(rule2)
  )
)
}\if{html}{\out{</div>}}

Or we can apply the same rule to multiple instruments:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{algos <- list(
  list( ## algo 1: algos[[1]]
    instruments = list("instr1", "instr2"),
    rules = list(rule1)
  )
)
}\if{html}{\out{</div>}}

Or multiple rules to a single instrument.:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{algos <- list(
  list( ## algo 1: algos[[1]]
    instruments = list("instr1"),
    rules = list(rule1, rule2)
  )
)
}\if{html}{\out{</div>}}
}
