% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{f_ewa}
\alias{f_ewa}
\title{Calculate Exponentially Weighted Average}
\usage{
f_ewa(x, lookback = NA)
}
\arguments{
\item{x}{vector}

\item{lookback}{Lookback window length as positive integer. If no \code{lookback}
is provided, the entire \eqn{x} vector will be used.}
}
\value{
Single exponentially weighted average value
}
\description{
Calculate EWA of vector \eqn{x} at time \eqn{t} based on a lookback window.
If the length of the lookback window is \eqn{L}, the range of the lookback
window is \eqn{[t-L, t-1]}.
}
