% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_ewa}
\alias{f_ewa}
\title{Calculate Exponentially Weighted Average}
\usage{
f_ewa(x, lambda = NA, lookback = NA)
}
\arguments{
\item{x}{Vector. Top to bottom: Oldest to newest.}

\item{lambda}{Smoothing parameter. If no lambda is provided, set lambda to  1 - (2 / (1 + L)).}

\item{lookback}{Lookback window length as positive integer. If no \code{lookback}
is provided, the entire \eqn{x} vector will be used.}
}
\value{
Single exponentially weighted average value
}
\description{
Calculate EWA of vector \eqn{x} at time \eqn{t} based on a lookback window.
If the length of the lookback window is \eqn{\lambda}, the range of the lookback
window is \eqn{[t-\lambda, t-1]}.
}
\details{
\deqn{E[X_t | X_{t-1}] = \frac{1}{\sum_{i=0}^{t-2}\lambda^i}\sum_{j=1}^{t-1}\lambda^{j-1} X_{t-j}}
}
\references{
Tsay: Analysis Of Financial Time Series
}
