% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_ewa_cor}
\alias{f_ewa_cor}
\title{Exponential Wrighted Correlation}
\usage{
f_ewa_cor(x, y, lookback, min_cor_mat)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}

\item{lookback}{Integer. Lookback window length. If no \code{lookback} is
provided, the entire \eqn{x} and \eqn{y} vectors will be used.}
}
\value{
Number
}
\description{
Exponential Wrighted Correlation
}
