% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_ewa_cov}
\alias{f_ewa_cov}
\title{Exponential Weighted Covariance}
\usage{
f_ewa_cov(x, y, lambda = NA, lookback = NA)
}
\arguments{
\item{x}{Vector.}

\item{y}{Vector.}

\item{lookback}{Integer. Lookback window length. If no \code{lookback} is
provided, the entire \eqn{x} and \eqn{y} vectors will be used.}
}
\value{
Number
}
\description{
Exponential Weighted Covariance
}
