% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{f_growth_rate_lvrg}
\alias{f_growth_rate_lvrg}
\title{Growth Rate With Leverage}
\usage{
f_growth_rate_lvrg(L, r, b)
}
\arguments{
\item{L}{Leverage factor.}

\item{r}{Return rate on investment.}

\item{b}{Borrowing rate.}
}
\value{
Growth rate as a decimal fraction.
}
\description{
Growth Rate With Leverage
}
