% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_inst_risk}
\alias{f_inst_risk}
\title{Instrument risk, annualized}
\usage{
f_inst_risk(prices, t, window_length = 25)
}
\arguments{
\item{prices}{A time series of prices in currency. Newest first.
Top to bottom: Newer to older.}

\item{t}{Time index.}

\item{window_length}{Window length. Includes all if \code{window_length} is NA.}
}
\value{
Standard deviation
}
\description{
Annualised standard deviation of returns in percentage terms.
\eqn{P_t}, price in currency at time \eqn{t}.
\eqn{n}, number of price observations.

\deqn{\overline{R} = \dfrac{1}{n} \sum_{t=2}^n \left(\dfrac{P_t}{P_{t-1}} - 1\right)}

\deqn{\text{Instrument Risk} = \sqrt{\dfrac{1}{n-1}\sum_{t=2}^n \left(\left(\dfrac{P_t}{P_{t-1}} - 1\right) - \overline{R}\right)^2}}

The square root of the number of business days per year is hard coded:

\code{sqrt(252) = 15.87451}

252 is the number of business days in a year.
(We could also round off to 16, and we would probably be fine.)
}
