% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_instrument_risk_target}
\alias{f_instrument_risk_target}
\title{Instrument Risk Target}
\usage{
f_instrument_risk_target(system_risk_target, IDM)
}
\arguments{
\item{system_risk_target}{Annualized volatility target in percentages}

\item{IDM}{Instrument Diversification Multiplier
\deqn{\text{IDM} = \frac{1}{\sqrt{W^{T}HW}}}
where \eqn{H} is the correlation matrix for returns, and \eqn{W} is the vector
of instrument weights summing to one.}
}
\value{
Decimal fraction (percentage divided by 100)
}
\description{
Instrument Risk Target
}
