% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_min_capital}
\alias{f_min_capital}
\title{Minimum capital}
\usage{
f_min_capital(min_exposure, inst_risk, instrument_risk_target)
}
\arguments{
\item{min_exposure}{Minimum exposure}

\item{inst_risk}{Instrument risk}

\item{instrument_risk_target}{Instrument risk target}
}
\description{
Calculate minimum capital required to trade the system.
}
