% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-rules.R
\name{f_moving_average}
\alias{f_moving_average}
\title{Moving Average}
\usage{
f_moving_average(prices, t = NA, window_length)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{t}{Time index.}

\item{window_length}{Window length.}
}
\value{
A single number. One step of a moving average.
}
\description{
Calculates average of n items prior to time t in price vector.
N (length of price vector) smaller than or equal to n is accepted.
This will not be the desired moving average, but will also not fail.

f_moving_average() doesn't actually calculate a "moving" average. It only
calculates one step of the moving average, i.e. the mean of the past n
prices.

If the length of the price vector is shorter than the window length, the
window length will be set to the length of the price vector, and a warning
will be given. The function will not fail or abort in this case.
}
