% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_notional_exposure}
\alias{f_notional_exposure}
\title{Notional exposure}
\usage{
f_notional_exposure(signal, instrument_risk_target, capital, instrument_risk)
}
\arguments{
\item{capital}{Trading capital in currency.}

\item{instrument_risk}{Instrument risk as decimal fraction.}

\item{risk_target}{Risk target as decimal fraction.}

\item{normalized_signal}{Input signal are normalized to an expected absolute
value of 1 as default.}

\item{signal_min}{Normalized signal will be capped at a minimum of
min_signal.}

\item{signal_max}{Normalized signal will be capped at a maximum of
max_signal.}
}
\description{
Calculate the notional exposure.
}
