% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas_accounting.R
\name{f_percentage_returns}
\alias{f_percentage_returns}
\title{Calculate Percentage Returns}
\usage{
f_percentage_returns(prices, t)
}
\arguments{
\item{prices}{A price vector containing at least the two prices we need for
calculating returns. Oldest first. Top to bottom: Older to newer. The last
observation is time t.}
}
\value{
Percentage as decimal fraction
}
\description{
Calculate percentage returns.
}
