% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_price_unit_vol}
\alias{f_price_unit_vol}
\title{Price Unit Volatility (instrument risk in price units)}
\usage{
f_price_unit_vol(price, inst_risk)
}
\arguments{
\item{price}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{inst_risk}{instrument risk.}
}
\description{
Price Unit Volatility (instrument risk in price units)
}
