% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{f_required_leverage_factor}
\alias{f_required_leverage_factor}
\title{Calculate Required Leverage Factor (RLF)}
\usage{
f_required_leverage_factor(instrument_risk_target, instrument_risk)
}
\arguments{
\item{instrument_risk_target}{Instrument volatility target}

\item{instrument_risk}{Instrument volatility}
}
\value{
A single number
}
\description{
Calculates the required leverage factor based on a risk target and the
volatility of an instrument in the same units.

The risk measures must be compatible. E.g. if instrument risk target is an
annual instrument volatility target in decimal percentages, and instrument
risk is the volatility annualized volatility of daily returns in decimal
percentages.

Note, the annual instrument volatility target can be calculated from a
system volatility target by dividing this by the Instrument Diversification
Diversifier.
}
