% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-basic.R
\name{f_returns_from_prices}
\alias{f_returns_from_prices}
\title{Calculate Net Returns Vector From Price Vector}
\usage{
f_returns_from_prices(prices)
}
\arguments{
\item{prices}{A vector of prices in currency. Oldest first. Top to bottom:
Older to newer. The last observation is time t.}
}
\value{
Percentage returns
}
\description{
Calculate Net Returns Vector From Price Vector
}
