% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-rules.R
\name{f_stop_loss_gap}
\alias{f_stop_loss_gap}
\title{Calculate Stop Loss Gap}
\usage{
f_stop_loss_gap(price_unit_vol, stop_loss_fraction)
}
\arguments{
\item{price_unit_vol}{Volatility of returns in price units.}

\item{stop_loss_fraction}{Stop loss fraction.}
}
\description{
Calculate stop loss gap. Stop loss only takes effect if the difference
between the current price and the stop loss level is bigger than the stop
loss gap.
}
