% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-rules.R
\name{f_stop_loss_level}
\alias{f_stop_loss_level}
\title{Stop loss level}
\usage{
f_stop_loss_level(hwm, lwm, stop_loss_gap, direction = 0, rnd = FALSE)
}
\arguments{
\item{hwm}{High Water Mark.}

\item{lwm}{Low Water Mark.}

\item{stop_loss_gap}{Stop loss gap.}

\item{direction}{Is current trade long or short? 1 for long, -1 for short.}

\item{rnd}{If TRUE, add small random amount to stop loss level. Negative if
short.}
}
\description{
Stop loss level
}
