% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{finalize_positions}
\alias{finalize_positions}
\title{Finalize Positions}
\usage{
finalize_positions(
  t,
  prices,
  final_target_pos_ccy,
  t_last_position_entry,
  latest_trade_direction,
  position_table,
  config
)
}
\arguments{
\item{t}{Time index.}

\item{prices}{Price vector.}

\item{final_target_pos_ccy}{Final target position size in currency of
account.}

\item{t_last_position_entry}{Time index of last position entry.}

\item{latest_trade_direction}{Latest trade direction.}

\item{position_table}{Position table.}

\item{config}{Config list.}
}
\value{
List of final column entries for row \code{t}
}
\description{
Calculate final position columns for row \code{t} in \code{position_table}.
}
\details{
A complete position table consists of \emph{standard columns}, \emph{final columns} and
position \emph{modifier output columns}.
}
