% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{get_position_weights}
\alias{get_position_weights}
\title{Get Position Weights}
\usage{
get_position_weights(t, position_tables, capital, prefix = "pre_")
}
\arguments{
\item{t}{Time index.}

\item{position_tables}{Position tables list.}

\item{account_table}{Account table.}
}
\value{
\eqn{t \times 1} vector of weights
}
\description{
Calculates the position weight for instrument \eqn{i} as
\deqn{w_{i, t} = \dfrac{N_{i,t} \cdot P_{i, t}}{C_t}}
where \eqn{N_{i,t}} is the number of contracts of instrument \eqn{i} at
time \eqn{t}, \eqn{P_{t, i}} is the price of instrument \eqn{i} at
time \eqn{t}, and \eqn{C_t} is the total capital of the system.
}
