% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_rule_names_by_parsed_algo}
\alias{get_rule_names_by_parsed_algo}
\title{Get Rule Names By Parsed Algo}
\usage{
get_rule_names_by_parsed_algo(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\value{
List of character strings.
}
\description{
Gets the rule name as a character string from each parsed algo and places the
names in a list in the same order as they appear in the algos list.

NOTE:
Since each algo now only accepts exactly one rule variation name, this in
effect does the same as \code{get_rule_variation_names_by_parsed_algo()}, just in
a more convoluted way. The rule name is now simply identical to the rule
function name (whereas earlier, a \emph{rule} would be a list of an arbitrary
number of rule variation names).

Rule names are the names of elements in the rule list for each algo, i.e. the
same as the rule variation names.
}
