% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_algos.R
\name{get_unique_rule_names_from_parsed_algos_list}
\alias{get_unique_rule_names_from_parsed_algos_list}
\title{Get Unique Rule Names From Expanded Algos List}
\usage{
get_unique_rule_names_from_parsed_algos_list(parsed_algos)
}
\arguments{
\item{parsed_algos}{List of expanded algos.}
}
\value{
List of character strings.
}
\description{
Get the unique rule names from an expanded algos list. Each rule is a
combination of one signal rule and optionally one stop loss rule.

NOTE:
This in effect does the same as
\code{get_unique_rule_variation_names_by_parsed_algo()}. This is because
\code{get_rule_names_by_parsed_algo()} in effect does the same as
\code{get_rule_variation_names_by_parsed_algo()}. See help file for
\code{get_rule_names_by_parsed_algo()}.

To get the names of all unique signal and stop loss rules, use
\code{get_unique_rule_variation_names_by_parsed_algo()}.
}
