% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas-risk.R
\name{instr_risk}
\alias{instr_risk}
\title{Instrument risk, annualized}
\usage{
instr_risk(prices, t, window_length = NA)
}
\arguments{
\item{prices}{A time series of prices in currency. Newest first.
Top to bottom: Newer to older.}

\item{t}{Time index.}

\item{window_length}{Window length.}
}
\description{
Standard deviations of returns.
}
