% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{load_rule_functions}
\alias{load_rule_functions}
\title{Load Rule Functions}
\usage{
load_rule_functions(parsed_algos)
}
\arguments{
\item{parsed_algos}{}
}
\description{
Load rule functions as specified by function names as character strings in
parsed algos into a list. The functions must be loaded into the enclosing
(typically global) environment.
}
