% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_multipliers.R
\name{m_block_same_direction_entry}
\alias{m_block_same_direction_entry}
\title{Block Entry In Same Direction}
\usage{
m_block_same_direction_entry(
  t,
  prices,
  direction,
  t_last_position_entry,
  position_table,
  mode = 1
)
}
\arguments{
\item{t}{Time index.}

\item{prices}{Price vector.}

\item{direction}{Direction. 1 for long, -1 for short, 0 for flat.}

\item{t_last_position_entry}{Time index of last position entry.}

\item{position_table}{Position table.}

\item{mode}{Mode.}
}
\description{
Position multiplier
}
