% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_position_modifier_list.R
\name{make_position_modifiers_list}
\alias{make_position_modifiers_list}
\title{Make List Of Position Modifiers From Position Modifiers List}
\usage{
make_position_modifiers_list(position_modifiers = list(), inst_names)
}
\arguments{
\item{position_modifiers}{User provided named list of position modifier
functions.}

\item{inst_names}{Vector of instrument names.}
}
\value{
Named list
}
\description{
Creates a named list. Values are all \code{NA} and the names are the names of
instruments (or any vector of character strings provided as input).
}
\details{
Initialize list of position modifiers. User may provide a named list where
the names are instrument names and values are position modifier functions.

\code{make_position_modifiers_list()} creates a named list. Each element corresponds
to an instrument, in the order instruments occur in the inst_data list. The
name of each element is the name of the instrument. User provided position
modifier functions will be assigned to the appropriate element in the list.

Any position modifier function for which the name doesn't match any
instrument in the system, will be ignored.

If no position modifier function is assigned to an instrument, the position
of that instrument will not be modified.

Position modifier functions must comply with these specifications:
\itemize{
\item The first parameter must be \code{t}, current time index. \code{t} is provided by
\code{update_position_table_row()}.
\item The second parameter must be \code{position_size_ccy}. \code{position_size_ccy} is
provided by \code{update_position_table_row()}.
\item The first two parameters are variable parameters. This means that they are
provided with data as the system is being updated. Parameters that are not
fixed are variable. I.e. any parameters that are not provided with a value
when running \code{make_system()}, is a variable parameter.
\item Any parameters that are provided with a value when running \code{make_system()},
is a fixed parameter.
}

The input format of the \code{position_modifiers} list provided to \code{make_system()}
is (example):

\if{html}{\out{<div class="sourceCode R">}}\preformatted{position_modifiers <- list(
  list(
    instruments = list("inst1", "inst2", "inst3"),
    modifier = list(
      "f1",
      f1,
     y1 = 10
   )
 )
)
}\if{html}{\out{</div>}}

This will be parsed into a list formated as

\if{html}{\out{<div class="sourceCode R">}}\preformatted{position_modifiers <- list(
  inst1 = list(
    modifier_name = "f1",
    modifier_function = f1,
    variable_params = list(
      x1 = "x1"
    ),
    fixed_params = list(
      y1 = 10
    )
  ),
  inst2 = list(
    modifier_name = "f1",
    modifier_function = f1,
    variable_params = list(
      x1 = "x1"
    ),
    fixed_params = list(
      y1 = 10
    )
  )
)
}\if{html}{\out{</div>}}

First element in each input \code{position_modifiers[[i]]$modifier} list is the
modifier name, second element is the modifier function definition itself.
Remaining elements are the fixed params.
Note: You may provide any name as modifier name. If the modifier
function is assigned to a variable, you may put this variable as the
function definition. So the function name doesn't have to match the
modifier name. The modifier name is just a label for your convenience.
Note: Variable params are not provided by user, i.e. do not appear
in modifier list.

If the list of data sets only contains \code{"inst1"} and \code{"inst2"}, then \code{"inst3"}
will be ignored when parsing the \code{position_modifiers} list.
}
