% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{make_system}
\alias{make_system}
\title{Make System}
\usage{
make_system(
  algos = list(),
  init_capital = 1e+05,
  risk_target = 0.12,
  risk_window_length = 25,
  stop_loss_fraction = 0.5,
  min_periods = 1,
  signal_min = -2,
  signal_max = 2,
  ...
)
}
\arguments{
\item{algos}{A list of algos.}

\item{init_capital}{Initial capital.}

\item{risk_window_length}{Risk window length.}

\item{stop_loss_fraction}{Stop loss fraction.}

\item{min_periods}{Minimum number of periods (rows) in the data sets.
Defaults to 1.}

\item{...}{}

\item{system_risk_target}{System risk target as decimal fraction (percent
divided by 100).}

\item{mode}{\code{sim} for \emph{simulation mode}, \code{live} for \emph{live mode}.}
}
\value{
A list containing
\itemize{
\item a list of instrument data sets as data frames. The instrument name is
assigned to each data frame as a comment. So we can get the data frame
name:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{comment(expanded_algos[[1]]$data)
[1] "instrument1"
}\if{html}{\out{</div>}}
\item a list of parsed algos, which in turn contains an instrument name and a
list of two trade rule functions (one for opening and on for closing).
\item a list \code{signal_tables}, each containing a data frame.
\item a list \code{position_table}, each containing a data frame.
\item a dataframe \code{system_account_table}.
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{system <- list(
  instrument_data <- list(
    data.frame(),
    data.frame(),
    ...,
    data.frame()
  )
  algos <- list(
    list(
      instrument = <instrument_name_1>,
      algo = list(
        <opening_rule_function>,
        <closing_rule_function>
      )
    ),
    list(
      instrument = <instrument_name_1>,
      algo = list(
        <opening_rule_function>,
        <closing_rule_function>
      )
    ),
    ...,
    list(
      instrument = <instrument_name_1>,
      algo = list(
        <opening_rule_function>,
        <closing_rule_function>
      )
    ),
  ),
  signal_tables <- list(
    data.frame(),
    data.frame(),
    ...,
    data.frame()
  ),
  position_tables <- list(
    data.frame(),
    data.frame(),
    ...,
    data.frame()
  ),
  account_table <- data.frame()
)
}\if{html}{\out{</div>}}
}
\description{
Make System
}
