% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{modify_position}
\alias{modify_position}
\title{Modify Position}
\usage{
modify_position(
  position_modifier,
  mod_variable_param_vals,
  position_multipliers,
  mul_variable_param_vals,
  position_size_ccy,
  ...
)
}
\arguments{
\item{position_size_ccy}{}
}
\description{
Modify Position
}
