% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_average}
\alias{moving_average}
\title{Simulate returns
Moving Average}
\usage{
moving_average(prices, n, t = NA)
}
\arguments{
\item{prices}{Vector.}

\item{n}{Window length.}

\item{t}{N (length of price vector) smaller than or equal to n is accepted.
This will not be the desired moving average, but will also not fail.}
}
\description{
Simulate returns
Moving Average
}
