% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_average_crossover}
\alias{moving_average_crossover}
\title{Moving Average Crossover
F13}
\usage{
moving_average_crossover(ma_fast, ma_slow, gap = 0)
}
\arguments{
\item{ma_fast}{Fast moving average}

\item{ma_slow}{Slow moving average
#' @param gap

Returns TRUE when ma_fast > ma_slow.
1 indicates uptrend ie. go long.
-1 indicates downtrend ie. go short.}
}
\description{
Moving Average Crossover
F13
}
