% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{moving_average_vector}
\alias{moving_average_vector}
\title{Moving Average vector}
\usage{
moving_average_vector(prices, n)
}
\arguments{
\item{prices}{Vector.}

\item{n}{Window length.

N (length of price vector) smaller than or equal to n is accepted.
This will not be the desired moving average, but will also not fail.}
}
\description{
Moving Average vector
}
