% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_multipliers.R
\name{o_limit_pf_risk}
\alias{o_limit_pf_risk}
\title{Limit Portfolio Risk}
\usage{
o_limit_pf_risk(t, position_tables, max_risk = NA, capital, cov_method)
}
\arguments{
\item{t}{Time index.}

\item{position_tables}{List of position tables.}

\item{max_risk}{Maximum risk level.}

\item{capital}{Capital.}

\item{cor_method}{Method of correlation calculation as character string.
\itemize{
\item \code{"Pearson"}
\item \code{"ewa"} Exponentially Weighted Correlation. @seealso \code{\link[=f_ewa_cor_mat]{f_ewa_cor_mat()}}
\itemize{
\item Additional parameters: \code{lambda} (number), \code{lookback} (integer)
}
}}
}
\value{
List of multiplier output. The first element must be a multiplier
value in \eqn{[0, 1]}. Additional elements are optional.
}
\description{
Portfolio risk limiting multiplier.
}
