% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_multipliers.R
\name{o_limit_pf_shock_risk}
\alias{o_limit_pf_shock_risk}
\title{Limit Shock Portfolio Risk}
\usage{
o_limit_pf_shock_risk(
  t,
  position_tables,
  max_risk = NA,
  sd_percentile,
  capital,
  sd_window_length,
  sd_method = "unbiased",
  cor_method = "Pearson",
  ...
)
}
\arguments{
\item{t}{Time index.}

\item{position_tables}{Position tables.}

\item{max_risk}{Maximum risk. Risk limit.}

\item{sd_percentile}{Shock percentile. Worst percentile of standard deviations.}

\item{capital}{Total system capital.}

\item{sd_window_length}{Standard deviation window length.}

\item{sd_method}{Standard deviation calculation method.}

\item{cor_method}{Correlation calculation method.}

\item{...}{}
}
\value{
List of multiplier output. The first element is the multiplier
value in \eqn{[0, 1]}. The second element is \code{shock_pf_risk}.
}
\description{
Limit Shock Portfolio Risk
}
