% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execution.R
\name{position_size_units}
\alias{position_size_units}
\title{Position size in units}
\usage{
position_size_units(price, risk_target, capital, instrument_risk)
}
\arguments{
\item{price}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{risk_target}{Risk target.}

\item{capital}{Capital.}

\item{instrument_risk}{Instrument risk.}
}
\value{
Position size in units.
}
\description{
Position size in units
}
