% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{price_unit_vol}
\alias{price_unit_vol}
\title{Price unit volatility (instrument risk in price units)}
\usage{
price_unit_vol(instr_risk, price)
}
\arguments{
\item{price}{LT F22}
}
\description{
Price unit volatility (instrument risk in price units)
}
