% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_functions.R
\name{r_stop_loss}
\alias{r_stop_loss}
\title{Stop loss rule}
\usage{
r_stop_loss(
  t,
  prices,
  instrument_risk,
  stop_loss_fraction,
  t_last_position_entry,
  direction,
  rnd = FALSE
)
}
\arguments{
\item{t}{Time index.}

\item{prices}{A vector of prices in currency. Oldest first. Top to bottom:
Older to newer.}

\item{instrument_risk}{Instrument risk at time t-1.}

\item{stop_loss_fraction}{Stop loss fraction.}

\item{direction}{Was current trade long or short at time t-1? 1 for long, -1
for short.}

\item{rnd}{If TRUE, add small random amount to stop loss level. Negative if
short.}

\item{t_trade_entry}{Time index of the time when current trade was entered.}
}
\value{
A named list containing:
\itemize{
\item \code{signal} Trade on/off indicator: 1 or 0
\item \code{stop_loss_gap} Stop loss gap
}
}
\description{
Return 0 if price level breached stop loss threshold. Return 1 if stop loss
threshold was not breached.

0 indicates that an open position should be exited. 1 indicates that an open
position should remain open. If a position is closed, and the stop loss
returns value 1, the position will remain closed.

Note that \code{instrument_risk}, \code{t_last_position_entry} and \code{direction} are
calculated at time \code{t - 1}. If we want these parameters at time \code{t}, we
need to apply the stop loss rule after \code{update_position_table_row}.
}
