% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{stop_loss_rule}
\alias{stop_loss_rule}
\title{Stop loss rule}
\usage{
stop_loss_rule(
  prices,
  t = NA,
  instrument_risk,
  stop_loss_fraction,
  t_trade_open,
  direction,
  rnd = FALSE
)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{t}{Time index.}

\item{instrument_risk}{Instrument risk.}

\item{stop_loss_fraction}{Stop loss fraction.}

\item{t_trade_open}{Time index of the time when trade is opened.}

\item{direction}{Is current trade long or short? 1 for long, -1 for short.}

\item{rnd}{If TRUE, add small random amount to stop loss level. Negative if
short.}
}
\value{
List:
\itemize{
\item Open/close indicator: 0 or 1
\item Stop loss gap
}
}
\description{
Return 0 if price level breached stop loss threshold. Return 1 if stop loss
threshold was not breached.

0 indicates that an open position should be exited. 1 indicates that an open
position should remain open.
}
