% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_sim.R
\name{trade_sim}
\alias{trade_sim}
\title{Execute trades in simulation mode}
\usage{
trade_sim(
  prices,
  init_capital = 1e+05,
  n_fast = 16,
  n_slow = 64,
  risk_target = 0.12,
  risk_window_length = 25,
  stop_loss_fraction = 0.5
)
}
\arguments{
\item{prices}{A vector of prices in currency. Newest first. Top to bottom:
Newer to older.}

\item{init_capital}{Initial capital.}

\item{n_fast}{n for fast moving average.}

\item{n_slow}{n for slow moving average.}

\item{risk_target}{Risk target.}

\item{risk_window_length}{Risk window length.}

\item{stop_loss_fraction}{Stop loss fraction.}
}
\value{
A list of trades data and accounts data.
}
\description{
Generates a dataframe of simulated trades from time series of prices.
}
