% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_price}
\alias{update_price}
\title{Update Price}
\usage{
update_price(inst_data, algo, t)
}
\arguments{
\item{inst_data}{List of unique data frames.}

\item{algo}{Algo list for a single expanded algo.}

\item{t}{Time index.}
}
\value{
Single price
}
\description{
Get latest price from data set in trade system list according to algo.
}
