% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_frwk.R
\name{update_signal_table_row}
\alias{update_signal_table_row}
\title{Update Signal Table Row}
\usage{
update_signal_table_row(
  t,
  inst_data,
  algo,
  signal_table,
  signal_weight,
  signal_normalization_factor,
  position_table,
  config
)
}
\arguments{
\item{t}{Time index.}

\item{algo}{Algo list for a single expanded algo.}

\item{signal_table}{Signal table data frame.}

\item{signal_weight}{Signal weight.}

\item{config}{List of trade system configuration values.}
}
\value{
List of signal table row values
}
\description{
Update Signal Table Row
}
